cleaning up the spin2 code.... now to do spin1 and update P1 outline code
