+++ spin2/package.json
@@ -3,7 +3,7 @@
   "displayName": "Spin2",
   "description": "Spin2/Pasm2 Syntax/Semantic Highlighting w/Code Outline and Custom tabbing support",
   "icon": "images/Propeller.ico",
-  "version": "1.8.1",
+  "version": "1.8.2",
   "publisher": "IronSheepProductionsLLC",
   "repository": {
     "type": "git",
