+++ spin2/src/spin2.semanticAndOutline.ts
@@ -4,7 +4,7 @@
 //import { deepStrictEqual } from "assert";
 
 import { semanticConfiguration, reloadSemanticConfiguration } from "./spin2.semantic.configuration";
-import { ParseUtils } from "./spin2.utils";
+import { ParseUtils, eDebugDisplayType, debugTypeForDisplay } from "./spin2.utils";
 
 import * as vscode from "vscode";
 
@@ -491,19 +491,6 @@ interface IFilteredStrings {
   lineParts: string[];
 }
 
-enum eDebugDisplayType {
-  Unknown = 0,
-  ddtLogic,
-  ddtScope,
-  ddtScopeXY,
-  ddtFFT,
-  ddtSpectro,
-  ddtPlot,
-  ddtTerm,
-  ddtBitmap,
-  ddtMidi,
-}
-
 enum eSpin2Directive {
   Unknown = 0,
   s2dDebugDisplayForLine,
@@ -515,19 +502,9 @@ interface ISpin2Directive {
 }
 
 // map of display-type to etype'
-const debugTypeForDisplay = new Map<string, eDebugDisplayType>([
-  ["logic", eDebugDisplayType.ddtLogic],
-  ["scope", eDebugDisplayType.ddtScope],
-  ["scope_xy", eDebugDisplayType.ddtScopeXY],
-  ["fft", eDebugDisplayType.ddtFFT],
-  ["spectro", eDebugDisplayType.ddtSpectro],
-  ["plot", eDebugDisplayType.ddtPlot],
-  ["term", eDebugDisplayType.ddtTerm],
-  ["bitmap", eDebugDisplayType.ddtBitmap],
-  ["midi", eDebugDisplayType.ddtMidi],
-]);
-
 export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSemanticTokensProvider {
+  private parseUtils = new ParseUtils();
+
   async provideDocumentSemanticTokens(document: vscode.TextDocument, cancelToken: vscode.CancellationToken): Promise<vscode.SemanticTokens> {
     // SEE https://www.codota.com/code/javascript/functions/vscode/CancellationToken/isCancellationRequested
     if (cancelToken) {
@@ -622,7 +599,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     for (let i = 0; i < lines.length; i++) {
       const line = lines[i];
       const trimmedLine = line.trim();
-      const trimmedNonCommentLine: string = this._getNonCommentLineRemainder(0, line);
+      const trimmedNonCommentLine: string = this.parseUtils.getNonCommentLineRemainder(0, line);
       const sectionStatus = this._isSectionStartLine(line);
       const lineParts: string[] = trimmedNonCommentLine.split(/[ \t]/);
       if (currState == eParseState.inMultiLineComment) {
@@ -643,7 +620,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
         }
         //  DO NOTHING Let Syntax highlighting do this
         continue;
-      } else if (this._isFlexspinPreprocessorDirective(lineParts[0])) {
+      } else if (this.parseUtils.isFlexspinPreprocessorDirective(lineParts[0])) {
         this._getPreProcessor_Declaration(0, i + 1, line);
         continue;
       } else if (sectionStatus.isSectionStart) {
@@ -664,7 +641,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           // process a class(static) variable line
           if (trimmedNonCommentLine.length > 6 && trimmedNonCommentLine.toUpperCase().includes("ORG")) {
             // ORG, ORGF, ORGH
-            const nonStringLine: string = this._removeDoubleQuotedStrings(trimmedNonCommentLine);
+            const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(trimmedNonCommentLine);
             if (nonStringLine.toUpperCase().includes("ORG")) {
               this._logPASM("- (" + (i + 1) + "): pre-scan DAT line trimmedLine=[" + trimmedLine + "] now Dat PASM");
               prePasmState = currState;
@@ -729,7 +706,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           if (trimmedLine.length > 6) {
             if (trimmedLine.toUpperCase().includes("ORG")) {
               // ORG, ORGF, ORGH
-              const nonStringLine: string = this._removeDoubleQuotedStrings(trimmedLine);
+              const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(trimmedLine);
               if (nonStringLine.toUpperCase().includes("ORG")) {
                 this._logPASM("- (" + (i + 1) + "): pre-scan DAT line trimmedLine=[" + trimmedLine + "] now Dat PASM");
                 prePasmState = currState;
@@ -841,7 +818,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           currState = priorState;
         }
         //  DO NOTHING Let Syntax highlighting do this
-      } else if (this._isFlexspinPreprocessorDirective(lineParts[0])) {
+      } else if (this.parseUtils.isFlexspinPreprocessorDirective(lineParts[0])) {
         const partialTokenSet: IParsedToken[] = this._reportPreProcessorLine(i, 0, line);
         partialTokenSet.forEach((newToken) => {
           this._logPreProc("=> PreProc: " + this._tokenString(newToken, line));
@@ -877,7 +854,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           // process a possible constant use on the CON line itself!
           if (line.length > 3) {
             if (trimmedLine.length > 6) {
-              const nonCommentLineRemainder: string = this._getNonCommentLineRemainder(0, trimmedLine);
+              const nonCommentLineRemainder: string = this.parseUtils.getNonCommentLineRemainder(0, trimmedLine);
               let orgStr: string = "ORGH";
               let orgOffset: number = nonCommentLineRemainder.toUpperCase().indexOf(orgStr); // ORGH
               if (orgOffset == -1) {
@@ -890,7 +867,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               }
               if (orgOffset != -1) {
                 // let's double check this is NOT in quoted string
-                const nonStringLine: string = this._removeDoubleQuotedStrings(nonCommentLineRemainder);
+                const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(nonCommentLineRemainder);
                 orgOffset = nonStringLine.toUpperCase().indexOf(orgStr); // ORG, ORGF, ORGH
               }
               if (orgOffset != -1) {
@@ -980,7 +957,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
         // process a line in a data section
         if (trimmedLine.length > 0) {
           this._logDAT("- process DAT line(" + (i + 1) + "): trimmedLine=[" + trimmedLine + "]");
-          const nonCommentLineRemainder: string = this._getNonCommentLineRemainder(0, trimmedLine);
+          const nonCommentLineRemainder: string = this.parseUtils.getNonCommentLineRemainder(0, trimmedLine);
           let orgStr: string = "ORGH";
           let orgOffset: number = nonCommentLineRemainder.toUpperCase().indexOf(orgStr); // ORGH
           if (orgOffset == -1) {
@@ -993,7 +970,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           }
           if (orgOffset != -1) {
             // let's double check this is NOT in quoted string
-            const nonStringLine: string = this._removeDoubleQuotedStrings(nonCommentLineRemainder);
+            const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(nonCommentLineRemainder);
             orgOffset = nonStringLine.toUpperCase().indexOf(orgStr); // ORG, ORGF, ORGH
           }
           if (orgOffset != -1) {
@@ -1105,7 +1082,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     if (line.toLowerCase().indexOf("{-* vscode-spin2:") != -1) {
       this._logMessage("- _getSpin2_Directive: ofs:" + startingOffset + ", [" + line + "](" + lineNumber + ")");
       // have possible directive
-      let currentOffset: number = this._skipWhite(line, startingOffset);
+      let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
       // get line parts - we only care about first one
       let lineParts: string[] = line
         .substring(currentOffset)
@@ -1124,14 +1101,14 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
 
   private _getPreProcessor_Declaration(startingOffset: number, lineNbr: number, line: string): void {
     if (this.configuration.highlightFlexspin) {
-      let currentOffset: number = this._skipWhite(line, startingOffset);
-      const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+      let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+      const nonCommentConstantLine = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
       // get line parts - we only care about first one
       const lineParts: string[] = nonCommentConstantLine.split(/[ \t=]/);
       this._logPreProc("  - ln:" + lineNbr + " GetPreProcDecl lineParts=[" + lineParts + "]");
       const directive: string = lineParts[0];
       const symbolName: string | undefined = lineParts.length > 1 ? lineParts[1] : undefined;
-      if (this._isFlexspinPreprocessorDirective(directive)) {
+      if (this.parseUtils.isFlexspinPreprocessorDirective(directive)) {
         // check a valid preprocessor line for a declaration
         if (symbolName != undefined && directive.toLowerCase() == "#define") {
           this._logPreProc("  -- new PreProc Symbol=[" + symbolName + "]");
@@ -1147,8 +1124,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     //
     if (line.substr(startingOffset).length > 1) {
       //skip Past Whitespace
-      let currentOffset: number = this._skipWhite(line, startingOffset);
-      const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+      let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+      const nonCommentConstantLine = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
       this._logCON("  - ln:" + lineNbr + " GetCONDecl nonCommentConstantLine=[" + nonCommentConstantLine + "]");
 
       const haveEnumDeclaration: boolean = nonCommentConstantLine.startsWith("#");
@@ -1179,8 +1156,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           //this._logCON('  -- lineParts=[' + lineParts + ']');
           for (let index = 0; index < lineParts.length; index++) {
             let enumConstant: string = lineParts[index];
-            // use _isDebugInvocation to filter out use of debug invocation command from constant def'
-            if (this._isDebugInvocation(enumConstant)) {
+            // use parseUtils.isDebugInvocation to filter out use of debug invocation command from constant def'
+            if (this.parseUtils.isDebugInvocation(enumConstant)) {
               continue; // yep this is not a constant
             } else {
               // our enum name can have a step offset
@@ -1204,10 +1181,10 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // HAVE    bGammaEnable        BYTE   TRUE               ' comment
     //         didShow             byte   FALSE[256]
     //                             byte   FALSE[256]
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const dataDeclNonCommentStr = this._getNonCommentLineRemainder(currentOffset, line);
-    let lineParts: string[] = this._getNonWhiteLineParts(dataDeclNonCommentStr);
+    const dataDeclNonCommentStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    let lineParts: string[] = this.parseUtils.getNonWhiteLineParts(dataDeclNonCommentStr);
     this._logDAT("- GetDatDecl lineParts=[" + lineParts + "](" + lineParts.length + ")");
     let haveMoreThanDat: boolean = lineParts.length > 1 && lineParts[0].toUpperCase() == "DAT";
     if (haveMoreThanDat || lineParts[0].toUpperCase() != "DAT") {
@@ -1220,8 +1197,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
         typeIndex = 2;
         maxParts = 3;
       }
-      let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[nameIndex]);
-      const isDataDeclarationLine: boolean = lineParts.length > maxParts - 1 && haveLabel && this._isDatStorageType(lineParts[typeIndex]) ? true : false;
+      let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[nameIndex]);
+      const isDataDeclarationLine: boolean = lineParts.length > maxParts - 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[typeIndex]) ? true : false;
       let lblFlag: string = haveLabel ? "T" : "F";
       let dataDeclFlag: string = isDataDeclarationLine ? "T" : "F";
       this._logDAT("- GetDatDecl lineParts=[" + lineParts + "](" + lineParts.length + ") label=" + lblFlag + ", daDecl=" + dataDeclFlag);
@@ -1245,17 +1222,17 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _getDAT_PasmDeclaration(startingOffset: number, line: string): void {
     // HAVE    bGammaEnable        BYTE   TRUE               ' comment
     //         didShow             byte   FALSE[256]
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const datPasmRHSStr = this._getNonCommentLineRemainder(currentOffset, line);
-    const lineParts: string[] = this._getNonWhiteLineParts(datPasmRHSStr);
+    const datPasmRHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    const lineParts: string[] = this.parseUtils.getNonWhiteLineParts(datPasmRHSStr);
     //this._logPASM('- GetDATPasmDecl lineParts=[' + lineParts + ']');
     // handle name in 1 column
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
-      if (!this._isReservedPasmSymbols(labelName) && !labelName.toUpperCase().startsWith("IF_") && !labelName.toUpperCase().startsWith("_RET_")) {
+      if (!this.parseUtils.isReservedPasmSymbols(labelName) && !labelName.toUpperCase().startsWith("IF_") && !labelName.toUpperCase().startsWith("_RET_")) {
         // org in first column is not label name, nor is if_ conditional
         const labelType: string = isDataDeclarationLine ? "variable" : "label";
         var labelModifiers: string[] = [];
@@ -1273,8 +1250,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     //  -or-   segments[7]     : "isp_hub75_segment"
     //
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const remainingNonCommentLineStr: string = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const remainingNonCommentLineStr: string = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     //this._logOBJ('- RptObjDecl remainingNonCommentLineStr=[' + remainingNonCommentLineStr + ']');
     const remainingLength: number = remainingNonCommentLineStr.length;
     if (remainingLength > 0) {
@@ -1295,7 +1272,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     //this._logSPIN("  -- GLBL GetMethodDecl methodType=[" + methodType + "], matchIdx(" + matchIdx + "), isPrivate(" + isPrivate + ")");
 
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const startNameOffset = currentOffset;
     // find open paren
     currentOffset = line.indexOf("(", currentOffset);
@@ -1312,14 +1289,14 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _getSPIN_PasmDeclaration(startingOffset: number, line: string): void {
     // HAVE    next8SLine ' or .nextLine in col 0
     //         nPhysLineIdx        long    0
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const inLinePasmRHSStr = this._getNonCommentLineRemainder(currentOffset, line);
-    const lineParts: string[] = this._getNonWhiteLineParts(inLinePasmRHSStr);
+    const inLinePasmRHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    const lineParts: string[] = this.parseUtils.getNonWhiteLineParts(inLinePasmRHSStr);
     //this._logPASM('- GetInLinePasmDecl lineParts=[' + lineParts + ']');
     // handle name in 1 column
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
       const labelType: string = isDataDeclarationLine ? "variable" : "label";
@@ -1336,12 +1313,12 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // HAVE    long    demoPausePeriod   ' comment
     //
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const remainingNonCommentLineStr: string = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const remainingNonCommentLineStr: string = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     this._logVAR("- GetVarDecl remainingNonCommentLineStr=[" + remainingNonCommentLineStr + "]");
     const isMultiDeclaration: boolean = remainingNonCommentLineStr.includes(",");
-    let lineParts: string[] = this._getNonWhiteDataInitLineParts(remainingNonCommentLineStr);
-    const hasGoodType: boolean = this._isStorageType(lineParts[0]);
+    let lineParts: string[] = this.parseUtils.getNonWhiteDataInitLineParts(remainingNonCommentLineStr);
+    const hasGoodType: boolean = this.parseUtils.isStorageType(lineParts[0]);
     this._logVAR("  -- lineParts=[" + lineParts + "]");
     let nameSet: string[] = [];
     if (hasGoodType && lineParts.length > 1) {
@@ -1378,18 +1355,18 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // locate and collect debug() display user names and types
     //
     // HAVE    debug(`{displayType} {displayName} ......)            ' comment
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
-    const datPasmStatementStr = this._getNonDocCommentLineRemainder(currentOffset, line);
+    const datPasmStatementStr = this.parseUtils.getNonDocCommentLineRemainder(currentOffset, line);
     if (datPasmStatementStr.toLowerCase().startsWith("debug(`")) {
-      const lineParts: string[] = this._getDebugNonWhiteLineParts(datPasmStatementStr);
+      const lineParts: string[] = this.parseUtils.getDebugNonWhiteLineParts(datPasmStatementStr);
       //this._logDEBUG('  -- debug(...) lineParts=[' + lineParts + ']');
       if (lineParts.length >= 3) {
         const displayType: string = lineParts[1];
         if (displayType.startsWith("`")) {
           const newDisplayType: string = displayType.substring(1, displayType.length);
           //this._logDEBUG('  --- debug(...) newDisplayType=[' + newDisplayType + ']');
-          if (this._isDebugDisplayType(newDisplayType)) {
+          if (this.parseUtils.isDebugDisplayType(newDisplayType)) {
             const newDisplayName: string = lineParts[2];
             //this._logDEBUG('  --- debug(...) newDisplayType=[' + newDisplayType + '], newDisplayName=[' + newDisplayName + ']');
             this._setUserDebugDisplay(newDisplayType, newDisplayName);
@@ -1402,15 +1379,15 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportPreProcessorLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const nonCommentConstantLine = this._getNonCommentLineRemainder(currentOffset, line);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const nonCommentConstantLine = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
     // get line parts - we only care about first one
     const lineParts: string[] = nonCommentConstantLine.split(/[ \t=]/);
     this._logPreProc("  - ln:" + lineNumber + " reportPreProc lineParts=[" + lineParts + "]");
     const directive: string = lineParts[0];
     const symbolName: string | undefined = lineParts.length > 1 ? lineParts[1] : undefined;
     if (this.configuration.highlightFlexspin) {
-      if (this._isFlexspinPreprocessorDirective(directive)) {
+      if (this.parseUtils.isFlexspinPreprocessorDirective(directive)) {
         // record the directive
         tokenSet.push({
           line: lineNumber,
@@ -1443,7 +1420,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               ptTokenType: referenceDetails.type,
               ptTokenModifiers: updatedModificationSet,
             });
-          } else if (this._isFlexspinReservedWord(symbolName)) {
+          } else if (this.parseUtils.isFlexspinReservedWord(symbolName)) {
             // record a constant reference
             tokenSet.push({
               line: lineNumber,
@@ -1483,7 +1460,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportCON_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const nonCommentConstantLine = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
 
     const haveEnumDeclaration: boolean = nonCommentConstantLine.startsWith("#");
@@ -1539,7 +1516,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           const assignmentRHSStr = conDeclarationLine.substring(fistEqualOffset + 1).trim();
           currentOffset = line.indexOf(assignmentRHSStr); // skip to RHS of assignment
           this._logCON("  -- CON assignmentRHSStr=[" + assignmentRHSStr + "]");
-          const possNames: string[] = this._getNonWhiteCONLineParts(assignmentRHSStr);
+          const possNames: string[] = this.parseUtils.getNonWhiteCONLineParts(assignmentRHSStr);
           this._logCON("  -- possNames=[" + possNames + "]");
           for (let index = 0; index < possNames.length; index++) {
             const possibleName = possNames[index];
@@ -1576,11 +1553,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 });
               } else {
                 if (
-                  !this._isSpinReservedWord(namePart) &&
-                  !this._isBuiltinReservedWord(namePart) &&
-                  !this._isDebugMethod(namePart) &&
-                  !this._isDebugSymbol(namePart) &&
-                  !this._isUnaryOperator(namePart)
+                  !this.parseUtils.isSpinReservedWord(namePart) &&
+                  !this.parseUtils.isBuiltinReservedWord(namePart) &&
+                  !this.parseUtils.isDebugMethod(namePart) &&
+                  !this.parseUtils.isDebugSymbol(namePart) &&
+                  !this.parseUtils.isUnaryOperator(namePart)
                 ) {
                   this._logCON("  --  CON MISSING name=[" + namePart + "]");
                   tokenSet.push({
@@ -1643,7 +1620,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 });
               }
             }
-            if (enumConstant.substr(0, 1).match(/[a-zA-Z_]/) && !this._isDebugInvocation(enumConstant)) {
+            if (enumConstant.substr(0, 1).match(/[a-zA-Z_]/) && !this.parseUtils.isDebugInvocation(enumConstant)) {
               this._logCON("  -- B GLBL enumConstant=[" + enumConstant + "]");
               // our enum name can have a step offset
               const nameOffset = line.indexOf(enumConstant, currentOffset);
@@ -1666,14 +1643,14 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportDAT_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const dataDeclNonCommentStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
-    let lineParts: string[] = this._getNonWhiteLineParts(dataDeclNonCommentStr);
+    let lineParts: string[] = this.parseUtils.getNonWhiteLineParts(dataDeclNonCommentStr);
     this._logVAR("- rptDataDeclLn lineParts=[" + lineParts + "]");
     // remember this object name so we can annotate a call to it
     if (lineParts.length > 1) {
-      if (this._isStorageType(lineParts[0]) || lineParts[0].toUpperCase() == "FILE" || lineParts[0].toUpperCase() == "ORG") {
+      if (this.parseUtils.isStorageType(lineParts[0]) || lineParts[0].toUpperCase() == "FILE" || lineParts[0].toUpperCase() == "ORG") {
         // if we start with storage type (or FILE, or ORG), not name, process rest of line for symbols
         currentOffset = line.indexOf(lineParts[0], currentOffset);
         const allowLocalVarStatus: boolean = false;
@@ -1700,7 +1677,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             ptTokenType: referenceDetails.type,
             ptTokenModifiers: referenceDetails.modifiers,
           });
-        } else if (!this._isReservedPasmSymbols(newName)) {
+        } else if (!this.parseUtils.isReservedPasmSymbols(newName)) {
           this._logDAT("  --  DAT rDdl MISSING name=[" + newName + "]");
           tokenSet.push({
             line: lineNumber,
@@ -1723,7 +1700,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     } else if (lineParts.length == 1) {
       // handle name declaration only line: [name 'comment]
       let newName = lineParts[0];
-      if (!this._isAlignType(newName)) {
+      if (!this.parseUtils.isAlignType(newName)) {
         let referenceDetails: RememberedToken | undefined = undefined;
         if (this._isGlobalToken(newName)) {
           referenceDetails = this._getGlobalToken(newName);
@@ -1751,14 +1728,14 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     this._logDAT("- process ValueDeclaration line(" + (lineNumber + 1) + "): line=[" + line + "]");
 
     // process data declaration
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const dataValueInitStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     if (dataValueInitStr.length > 1) {
       if (showDebug) {
         this._logMessage("  -- reportDataValueInit dataValueInitStr=[" + dataValueInitStr + "]");
       }
-      let lineParts: string[] = this._getNonWhiteDataInitLineParts(dataValueInitStr);
-      const argumentStartIndex: number = this._isDatStorageType(lineParts[0]) ? 1 : 0;
+      let lineParts: string[] = this.parseUtils.getNonWhiteDataInitLineParts(dataValueInitStr);
+      const argumentStartIndex: number = this.parseUtils.isDatStorageType(lineParts[0]) ? 1 : 0;
       if (showDebug) {
         this._logMessage("  -- lineParts=[" + lineParts + "]");
       }
@@ -1816,13 +1793,13 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               });
             } else {
               if (
-                !this._isPasmReservedWord(namePart) &&
-                !this._isReservedPasmSymbols(namePart) &&
-                !this._isPasmInstruction(namePart) &&
-                !this._isDatNFileStorageType(namePart) &&
-                !this._isBinaryOperator(namePart) &&
-                !this._isUnaryOperator(namePart) &&
-                !this._isBuiltinReservedWord(namePart)
+                !this.parseUtils.isPasmReservedWord(namePart) &&
+                !this.parseUtils.isReservedPasmSymbols(namePart) &&
+                !this.parseUtils.isPasmInstruction(namePart) &&
+                !this.parseUtils.isDatNFileStorageType(namePart) &&
+                !this.parseUtils.isBinaryOperator(namePart) &&
+                !this.parseUtils.isUnaryOperator(namePart) &&
+                !this.parseUtils.isBuiltinReservedWord(namePart)
               ) {
                 if (showDebug) {
                   this._logMessage("  --  DAT rDvdc MISSING name=[" + namePart + "]");
@@ -1864,10 +1841,10 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportDAT_PasmCode(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const inLinePasmRHSStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
-    const lineParts: string[] = this._getNonWhitePasmLineParts(inLinePasmRHSStr);
+    const lineParts: string[] = this.parseUtils.getNonWhitePasmLineParts(inLinePasmRHSStr);
     currentOffset = line.indexOf(inLinePasmRHSStr, currentOffset);
     this._logPASM("  -- reportDATPasmDecl lineParts=[" + lineParts + "]");
     // handle name in 1 column
@@ -1879,8 +1856,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
         tokenSet.push(newToken);
       });
     }
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]);
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]);
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     // TODO: REWRITE this to handle "non-label" line with unknown op-code!
     if (haveLabel) {
       // process label/variable name - starting in column 0
@@ -1946,7 +1923,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               // skip empty operand
               continue;
             }
-            if (index == lineParts.length - 1 && this._isPasmConditional(argumentName)) {
+            if (index == lineParts.length - 1 && this.parseUtils.isPasmConditional(argumentName)) {
               // conditional flag-set spec.
               this._logPASM("  -- SKIP argumentName=[" + argumentName + "]");
               continue;
@@ -1989,14 +1966,14 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               } else {
                 // we use bIsDebugLine in next line so we don't flag debug() arguments!
                 if (
-                  !this._isPasmReservedWord(namePart) &&
-                  !this._isPasmInstruction(namePart) &&
-                  !this._isPasmConditional(namePart) &&
-                  !this._isBinaryOperator(namePart) &&
-                  !this._isBuiltinReservedWord(namePart) &&
-                  !this._isCoginitReservedSymbol(namePart) &&
-                  !this._isPasmModczOperand(namePart) &&
-                  !this._isDebugMethod(namePart) &&
+                  !this.parseUtils.isPasmReservedWord(namePart) &&
+                  !this.parseUtils.isPasmInstruction(namePart) &&
+                  !this.parseUtils.isPasmConditional(namePart) &&
+                  !this.parseUtils.isBinaryOperator(namePart) &&
+                  !this.parseUtils.isBuiltinReservedWord(namePart) &&
+                  !this.parseUtils.isCoginitReservedSymbol(namePart) &&
+                  !this.parseUtils.isPasmModczOperand(namePart) &&
+                  !this.parseUtils.isDebugMethod(namePart) &&
                   !bIsAlsoDebugLine
                 ) {
                   this._logPASM("  --  DAT Pasm MISSING name=[" + namePart + "](" + (nameOffset + 1) + ")");
@@ -2033,7 +2010,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       }
     } else {
       // process data declaration
-      if (this._isDatStorageType(lineParts[0])) {
+      if (this.parseUtils.isDatStorageType(lineParts[0])) {
         currentOffset = line.indexOf(lineParts[0], currentOffset);
       } else {
         // skip line part 0 length when searching for [1] name
@@ -2056,7 +2033,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     this.localTokens.clear();
     const isPrivate = methodType.indexOf("PRI") != -1;
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const spineDeclarationLHSStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     if (spineDeclarationLHSStr) {
     } // we don't use this string, we called this to record our rhs comment!
@@ -2212,11 +2189,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   });
                 } else {
                   if (
-                    !this._isSpinReservedWord(namedIndexPart) &&
-                    !this._isSpinBuiltinMethod(namedIndexPart) &&
-                    !this._isBuiltinReservedWord(namedIndexPart) &&
-                    !this._isDebugMethod(namedIndexPart) &&
-                    !this._isDebugSymbol(namedIndexPart)
+                    !this.parseUtils.isSpinReservedWord(namedIndexPart) &&
+                    !this.parseUtils.isSpinBuiltinMethod(namedIndexPart) &&
+                    !this.parseUtils.isBuiltinReservedWord(namedIndexPart) &&
+                    !this.parseUtils.isDebugMethod(namedIndexPart) &&
+                    !this.parseUtils.isDebugSymbol(namedIndexPart)
                   ) {
                     // we don't have name registered so just mark it
                     this._logSPIN("  --  SPIN MISSING varname=[" + namedIndexPart + "](" + (nameOffset + 1) + ")");
@@ -2247,7 +2224,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             this._setLocalToken(localName, new RememberedToken("variable", ["local"])); // TOKEN SET in _report()
           } else {
             // have modifier!
-            if (this._isStorageType(localName)) {
+            if (this.parseUtils.isStorageType(localName)) {
               tokenSet.push({
                 line: lineNumber,
                 startCharacter: nameOffset,
@@ -2255,7 +2232,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 ptTokenType: "storageType",
                 ptTokenModifiers: [],
               });
-            } else if (this._isAlignType(localName)) {
+            } else if (this.parseUtils.isAlignType(localName)) {
               tokenSet.push({
                 line: lineNumber,
                 startCharacter: nameOffset,
@@ -2275,7 +2252,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportSPIN_Code(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const nonCommentSpinLine = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     const remainingLength: number = nonCommentSpinLine.length;
     this._logCON("- reportSPIN nonCommentSpinLine=[" + nonCommentSpinLine + "] remainingLength=" + remainingLength);
@@ -2322,7 +2299,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               if (variableNamePart.substr(0, 1).match(/[a-zA-Z_]/)) {
                 if (variableNamePart.includes(".")) {
                   const varNameParts: string[] = variableNamePart.split(".");
-                  if (this._isDatStorageType(varNameParts[1])) {
+                  if (this.parseUtils.isDatStorageType(varNameParts[1])) {
                     variableNamePart = varNameParts[0]; // just use first part of name
                     /*
                                         // FIXME: UNDONE mark storage part correctly, yes, out-of-order
@@ -2339,7 +2316,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   }
                 }
                 this._logSPIN("  -- variableNamePart=[" + variableNamePart + "](" + (nameOffset + 1) + ")");
-                if (this._isStorageType(variableNamePart)) {
+                if (this.parseUtils.isStorageType(variableNamePart)) {
                   tokenSet.push({
                     line: lineNumber,
                     startCharacter: nameOffset,
@@ -2368,11 +2345,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                     });
                   } else {
                     if (
-                      !this._isSpinReservedWord(variableNamePart) &&
-                      !this._isBuiltinReservedWord(variableNamePart) &&
-                      !this._isDebugMethod(variableNamePart) &&
-                      !this._isDebugSymbol(variableNamePart) &&
-                      !this._isSpinBuiltinMethod(variableNamePart)
+                      !this.parseUtils.isSpinReservedWord(variableNamePart) &&
+                      !this.parseUtils.isBuiltinReservedWord(variableNamePart) &&
+                      !this.parseUtils.isDebugMethod(variableNamePart) &&
+                      !this.parseUtils.isDebugSymbol(variableNamePart) &&
+                      !this.parseUtils.isSpinBuiltinMethod(variableNamePart)
                     ) {
                       // we don't have name registered so just mark it
                       this._logSPIN("  --  SPIN MISSING varname=[" + variableNamePart + "](" + (nameOffset + 1) + ")");
@@ -2392,11 +2369,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             // have simple target name, no []
             let cleanedVariableName: string = variableName.replace(/[ \t\(\)]/, "");
             const nameOffset = line.indexOf(cleanedVariableName, currentOffset);
-            if (cleanedVariableName.substr(0, 1).match(/[a-zA-Z_]/) && !this._isStorageType(cleanedVariableName)) {
+            if (cleanedVariableName.substr(0, 1).match(/[a-zA-Z_]/) && !this.parseUtils.isStorageType(cleanedVariableName)) {
               this._logSPIN("  --  SPIN cleanedVariableName=[" + cleanedVariableName + "](" + (nameOffset + 1) + ")");
               if (cleanedVariableName.includes(".")) {
                 const varNameParts: string[] = cleanedVariableName.split(".");
-                if (this._isDatStorageType(varNameParts[1])) {
+                if (this.parseUtils.isDatStorageType(varNameParts[1])) {
                   cleanedVariableName = varNameParts[0]; // just use first part of name
                   /*
                                     // FIXME: UNDONE mark storage part correctly, yes, out-of-order
@@ -2442,11 +2419,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               } else {
                 // we don't have name registered so just mark it
                 if (
-                  !this._isSpinReservedWord(cleanedVariableName) &&
-                  !this._isSpinBuiltinMethod(cleanedVariableName) &&
-                  !this._isBuiltinReservedWord(cleanedVariableName) &&
-                  !this._isDebugMethod(cleanedVariableName) &&
-                  !this._isDebugSymbol(cleanedVariableName)
+                  !this.parseUtils.isSpinReservedWord(cleanedVariableName) &&
+                  !this.parseUtils.isSpinBuiltinMethod(cleanedVariableName) &&
+                  !this.parseUtils.isBuiltinReservedWord(cleanedVariableName) &&
+                  !this.parseUtils.isDebugMethod(cleanedVariableName) &&
+                  !this.parseUtils.isDebugSymbol(cleanedVariableName)
                 ) {
                   this._logSPIN("  --  SPIN MISSING cln name=[" + cleanedVariableName + "](" + (nameOffset + 1) + ")");
                   tokenSet.push({
@@ -2470,7 +2447,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       // -------------------------------------------
       const assignmentRHSStr: string = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
       currentOffset = line.indexOf(assignmentRHSStr, currentOffset);
-      const preCleanAssignmentRHSStr = this._getNonInlineCommentLine(assignmentRHSStr).replace("..", "  ");
+      const preCleanAssignmentRHSStr = this.parseUtils.getNonInlineCommentLine(assignmentRHSStr).replace("..", "  ");
       this._logSPIN("  -- SPIN assignmentRHSStr=[" + assignmentRHSStr + "]");
       const lineInfo: IFilteredStrings = this._getNonWhiteSpinLineParts(preCleanAssignmentRHSStr);
       let possNames: string[] = lineInfo.lineParts;
@@ -2528,7 +2505,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             });
           } else {
             // have unknown name!? is storage type spec?
-            if (this._isStorageType(namePart)) {
+            if (this.parseUtils.isStorageType(namePart)) {
               this._logSPIN("  --  SPIN RHS storageType=[" + namePart + "]");
               tokenSet.push({
                 line: lineNumber,
@@ -2540,13 +2517,13 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             }
             // we use bIsDebugLine in next line so we don't flag debug() arguments!
             else if (
-              !this._isSpinReservedWord(namePart) &&
-              !this._isSpinBuiltinMethod(namePart) &&
-              !this._isBuiltinReservedWord(namePart) &&
-              !this._isCoginitReservedSymbol(namePart) &&
-              !this._isDebugMethod(namePart) &&
+              !this.parseUtils.isSpinReservedWord(namePart) &&
+              !this.parseUtils.isSpinBuiltinMethod(namePart) &&
+              !this.parseUtils.isBuiltinReservedWord(namePart) &&
+              !this.parseUtils.isCoginitReservedSymbol(namePart) &&
+              !this.parseUtils.isDebugMethod(namePart) &&
               !bIsDebugLine &&
-              !this._isDebugInvocation(namePart)
+              !this.parseUtils.isDebugInvocation(namePart)
             ) {
               // NO DEBUG FOR ELSE, most of spin control elements come through here!
               //else {
@@ -2572,7 +2549,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               isMethod = true;
             }
             const constantPart: string = possibleNameSet[1];
-            if (this._isStorageType(constantPart)) {
+            if (this.parseUtils.isStorageType(constantPart)) {
               // FIXME: UNDONE remove when syntax see this correctly
               const nameOffset: number = line.indexOf(constantPart, currentOffset);
               this._logSPIN("  --  SPIN rhs storageType=[" + constantPart + "](" + (nameOffset + 1) + ")");
@@ -2618,10 +2595,10 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportSPIN_PasmCode(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const inLinePasmRHSStr = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
-    const lineParts: string[] = this._getNonWhitePasmLineParts(inLinePasmRHSStr);
+    const lineParts: string[] = this.parseUtils.getNonWhitePasmLineParts(inLinePasmRHSStr);
     this._logPASM("  -- reportInLinePasmDecl lineParts=[" + lineParts + "]");
     const bIsAlsoDebugLine: boolean = inLinePasmRHSStr.toLowerCase().indexOf("debug(") != -1 ? true : false;
     if (bIsAlsoDebugLine) {
@@ -2633,8 +2610,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     }
     // handle name in as first part of line...
     // (process label/variable name (but 'debug' of debug() is NOT a label!))
-    let haveLabel: boolean = this._isDatOrPasmLabel(lineParts[0]) && lineParts[0].toLowerCase() != "debug";
-    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this._isDatStorageType(lineParts[1]) ? true : false;
+    let haveLabel: boolean = this.parseUtils.isDatOrPasmLabel(lineParts[0]) && lineParts[0].toLowerCase() != "debug";
+    const isDataDeclarationLine: boolean = lineParts.length > 1 && haveLabel && this.parseUtils.isDatStorageType(lineParts[1]) ? true : false;
     if (haveLabel) {
       const labelName: string = lineParts[0];
       this._logPASM("  -- labelName=[" + labelName + "]");
@@ -2694,7 +2671,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 // skip empty operand
                 continue;
               }
-              if (index == lineParts.length - 1 && this._isPasmConditional(argumentName)) {
+              if (index == lineParts.length - 1 && this.parseUtils.isPasmConditional(argumentName)) {
                 // conditional flag-set spec.
                 this._logPASM("  -- SKIP argumentName=[" + argumentName + "]");
                 continue;
@@ -2737,7 +2714,12 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   // we don't have name registered so just mark it
                   if (namePart != ".") {
                     // odd special case!
-                    if (!this._isSpinReservedWord(namePart) && !this._isBuiltinReservedWord(namePart) && !this._isDebugMethod(namePart) && !this._isPasmModczOperand(namePart)) {
+                    if (
+                      !this.parseUtils.isSpinReservedWord(namePart) &&
+                      !this.parseUtils.isBuiltinReservedWord(namePart) &&
+                      !this.parseUtils.isDebugMethod(namePart) &&
+                      !this.parseUtils.isPasmModczOperand(namePart)
+                    ) {
                       // XYZZY
                       this._logPASM("  --  SPIN Pasm MISSING name=[" + namePart + "]");
                       tokenSet.push({
@@ -2747,7 +2729,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                         ptTokenType: "variable",
                         ptTokenModifiers: ["missingDeclaration"],
                       });
-                    } else if (this._isIllegalInlinePasmDirective(namePart)) {
+                    } else if (this.parseUtils.isIllegalInlinePasmDirective(namePart)) {
                       this._logPASM("  --  SPIN inlinePasm ERROR[CODE] illegal name=[" + namePart + "]");
                       tokenSet.push({
                         line: lineNumber,
@@ -2785,7 +2767,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
         const nameOrDirective: string = lineParts[0];
         // if this symbol is NOT a global token then it could be bad!
         if (!this._isKnownToken(nameOrDirective)) {
-          if (this._isIllegalInlinePasmDirective(nameOrDirective) || !this._isPasmInstruction(nameOrDirective)) {
+          if (this.parseUtils.isIllegalInlinePasmDirective(nameOrDirective) || !this.parseUtils.isPasmInstruction(nameOrDirective)) {
             this._logPASM("  --  SPIN inline-Pasm MISSING name=[" + nameOrDirective + "]");
             const nameOffset = line.indexOf(nameOrDirective, currentOffset);
             tokenSet.push({
@@ -2800,7 +2782,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       }
     } else {
       // process data declaration
-      if (this._isDatStorageType(lineParts[0])) {
+      if (this.parseUtils.isDatStorageType(lineParts[0])) {
         currentOffset = line.indexOf(lineParts[0], currentOffset);
       } else {
         currentOffset = line.indexOf(lineParts[1], currentOffset);
@@ -2818,7 +2800,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportOBJ_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const remainingNonCommentLineStr: string = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     //this._logOBJ('- RptObjDecl remainingNonCommentLineStr=[' + remainingNonCommentLineStr + ']');
     const remainingLength: number = remainingNonCommentLineStr.length;
@@ -2883,7 +2865,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   });
                 }
                 // we don't have name registered so just mark it
-                else if (!this._isSpinReservedWord(nameReference) && !this._isBuiltinReservedWord(nameReference) && !this._isDebugMethod(nameReference)) {
+                else if (!this.parseUtils.isSpinReservedWord(nameReference) && !this.parseUtils.isBuiltinReservedWord(nameReference) && !this.parseUtils.isDebugMethod(nameReference)) {
                   this._logOBJ("  --  OBJ MISSING name=[" + nameReference + "]");
                   const nameOffset = line.indexOf(nameReference, currentOffset);
                   tokenSet.push({
@@ -2906,15 +2888,15 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
   private _reportVAR_DeclarationLine(lineNumber: number, startingOffset: number, line: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     //skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     const remainingNonCommentLineStr: string = this._getNonCommentLineReturnComment(lineNumber, currentOffset, line, tokenSet);
     if (remainingNonCommentLineStr.length > 0) {
       // get line parts - we only care about first one
-      let lineParts: string[] = this._getCommaDelimitedNonWhiteLineParts(remainingNonCommentLineStr);
+      let lineParts: string[] = this.parseUtils.getCommaDelimitedNonWhiteLineParts(remainingNonCommentLineStr);
       this._logVAR("  -- rptVarDecl lineParts=[" + lineParts + "]");
       // remember this object name so we can annotate a call to it
       const isMultiDeclaration: boolean = remainingNonCommentLineStr.includes(",");
-      const hasStorageType: boolean = this._isStorageType(lineParts[0]);
+      const hasStorageType: boolean = this.parseUtils.isStorageType(lineParts[0]);
       if (lineParts.length > 1) {
         const startIndex: number = hasStorageType ? 1 : 0;
         for (let index = startIndex; index < lineParts.length; index++) {
@@ -2974,7 +2956,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                     });
                   } else {
                     // we don't have name registered so just mark it
-                    if (!this._isSpinReservedWord(namePart) && !this._isBuiltinReservedWord(namePart) && !this._isDebugMethod(namePart)) {
+                    if (!this.parseUtils.isSpinReservedWord(namePart) && !this.parseUtils.isBuiltinReservedWord(namePart) && !this.parseUtils.isDebugMethod(namePart)) {
                       this._logVAR("  --  VAR Add MISSING name=[" + namePart + "]");
                       tokenSet.push({
                         line: lineNumber,
@@ -3031,7 +3013,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     // debug(`lstr_(displayName, len) lutcolors `uhex_long_array_(image_address, lut_size))
     // debug(``#(letter) lutcolors `uhex_long_array_(image_address, lut_size))
     //
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     // get line parts - we only care about first one
     const debugStatementStr = this._getDebugStatement(currentOffset, line);
     this._logDEBUG(" -- rptDbg debugStatementStr=[" + debugStatementStr + "]");
@@ -3052,7 +3034,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
       tokenSet.push(newToken);
     }
     this._logDEBUG("-- DEBUG line(" + lineNumber + ") debugStatementStr=[" + debugStatementStr + "]");
-    let lineParts: string[] = this._getDebugNonWhiteLineParts(debugStatementStr);
+    let lineParts: string[] = this.parseUtils.getDebugNonWhiteLineParts(debugStatementStr);
     this._logDEBUG(" -- rptDbg lineParts=[" + lineParts + "]");
     if (lineParts.length > 0 && lineParts[0].toLowerCase() != "debug") {
       //this._logDEBUG(' -- rptDbg first name not debug! (label?) removing! lineParts[0]=[' + lineParts[0] + ']');
@@ -3070,7 +3052,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           displayTestName = displayTestName.toLowerCase().replace(/ \t/g, "");
           const isRuntimeNamed: boolean = displayTestName.startsWith("``") || displayTestName.startsWith("`zstr") || displayTestName.startsWith("`lstr");
           this._logDEBUG(" -- rptDbg displayTestName=[" + displayTestName + "], isRuntimeNamed=" + isRuntimeNamed);
-          let bHaveInstantiation = this._isDebugDisplayType(newDisplayType) && !isRuntimeNamed;
+          let bHaveInstantiation = this.parseUtils.isDebugDisplayType(newDisplayType) && !isRuntimeNamed;
           if (bHaveInstantiation) {
             this._logDEBUG("  -- rptDbg --- PROCESSING Instantiation");
             // -------------------------------------
@@ -3104,7 +3086,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             for (let idx = firstParamIdx; idx < lineParts.length; idx++) {
               const newParameter: string = lineParts[idx];
               symbolOffset = line.indexOf(newParameter, symbolOffset);
-              const bIsParameterName: boolean = this._isNameWithTypeInstantiation(newParameter, eDisplayType);
+              const bIsParameterName: boolean = this.parseUtils.isNameWithTypeInstantiation(newParameter, eDisplayType);
               if (bIsParameterName) {
                 this._logDEBUG("  -- rptDbg newParam=[" + newParameter + "]");
                 tokenSet.push({
@@ -3115,7 +3097,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                   ptTokenModifiers: ["reference", "defaultLibrary"],
                 });
               } else {
-                const bIsColorName: boolean = this._isDebugColorName(newParameter);
+                const bIsColorName: boolean = this.parseUtils.isDebugColorName(newParameter);
                 if (bIsColorName) {
                   this._logDEBUG("  -- rptDbg newColor=[" + newParameter + "]");
                   tokenSet.push({
@@ -3152,13 +3134,13 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                     const paramIsSymbolName: boolean = newParameter.substring(0, 1).match(/[a-zA-Z_]/) ? true : false;
                     if (
                       paramIsSymbolName &&
-                      !this._isDebugMethod(newParameter) &&
+                      !this.parseUtils.isDebugMethod(newParameter) &&
                       newParameter.indexOf("`") == -1 &&
-                      !this._isUnaryOperator(newParameter) &&
-                      !this._isBinaryOperator(newParameter) &&
-                      !this._isFloatConversion(newParameter) &&
-                      !this._isSpinBuiltinMethod(newParameter) &&
-                      !this._isBuiltinReservedWord(newParameter)
+                      !this.parseUtils.isUnaryOperator(newParameter) &&
+                      !this.parseUtils.isBinaryOperator(newParameter) &&
+                      !this.parseUtils.isFloatConversion(newParameter) &&
+                      !this.parseUtils.isSpinBuiltinMethod(newParameter) &&
+                      !this.parseUtils.isBuiltinReservedWord(newParameter)
                     ) {
                       this._logDEBUG("  -- rptDbg 1 unkParam=[" + newParameter + "]");
                       tokenSet.push({
@@ -3232,7 +3214,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
               let newParameter: string = "";
               for (let idx = firstParamIdx; idx < lineParts.length; idx++) {
                 newParameter = lineParts[idx];
-                if (newParameter.indexOf("'") != -1 || this._isStorageType(newParameter)) {
+                if (newParameter.indexOf("'") != -1 || this.parseUtils.isStorageType(newParameter)) {
                   symbolOffset += newParameter.length;
                   continue; // skip this name (it's part of a string!)
                 } else if (newParameter.indexOf("#") != -1) {
@@ -3241,7 +3223,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 }
                 symbolOffset = line.indexOf(newParameter, symbolOffset);
                 this._logDEBUG("  -- rptDbg ?check? [" + newParameter + "] symbolOffset=" + symbolOffset);
-                let bIsParameterName: boolean = this._isNameWithTypeFeed(newParameter, eDisplayType);
+                let bIsParameterName: boolean = this.parseUtils.isNameWithTypeFeed(newParameter, eDisplayType);
                 if (isRuntimeNamed && newParameter.toLowerCase() == "lutcolors") {
                   bIsParameterName = true;
                 }
@@ -3255,7 +3237,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                     ptTokenModifiers: ["reference", "defaultLibrary"],
                   });
                 } else {
-                  const bIsColorName: boolean = this._isDebugColorName(newParameter);
+                  const bIsColorName: boolean = this.parseUtils.isDebugColorName(newParameter);
                   if (bIsColorName) {
                     this._logDEBUG("  -- rptDbg newColor=[" + newParameter + "]");
                     tokenSet.push({
@@ -3292,13 +3274,13 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                       const paramIsSymbolName: boolean = newParameter.substring(0, 1).match(/[a-zA-Z_]/) ? true : false;
                       if (
                         paramIsSymbolName &&
-                        this._isDebugMethod(newParameter) == false &&
+                        this.parseUtils.isDebugMethod(newParameter) == false &&
                         newParameter.indexOf("`") == -1 &&
-                        !this._isUnaryOperator(newParameter) &&
-                        !this._isBinaryOperator(newParameter) &&
-                        !this._isFloatConversion(newParameter) &&
-                        !this._isSpinBuiltinMethod(newParameter) &&
-                        !this._isBuiltinReservedWord(newParameter)
+                        !this.parseUtils.isUnaryOperator(newParameter) &&
+                        !this.parseUtils.isBinaryOperator(newParameter) &&
+                        !this.parseUtils.isFloatConversion(newParameter) &&
+                        !this.parseUtils.isSpinBuiltinMethod(newParameter) &&
+                        !this.parseUtils.isBuiltinReservedWord(newParameter)
                       ) {
                         this._logDEBUG("  -- rptDbg 2 unkParam=[" + newParameter + "]"); // XYZZY LutColors
                         tokenSet.push({
@@ -3330,7 +3312,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
           let newParameter: string = "";
           for (let idx = firstParamIdx; idx < lineParts.length; idx++) {
             newParameter = lineParts[idx];
-            if (newParameter.toLowerCase() == "debug" || this._isStorageType(newParameter)) {
+            if (newParameter.toLowerCase() == "debug" || this.parseUtils.isStorageType(newParameter)) {
               continue;
             }
             symbolOffset = line.indexOf(newParameter, symbolOffset); // walk this past each
@@ -3368,12 +3350,12 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
                 const paramIsSymbolName: boolean = newParameter.substring(0, 1).match(/[a-zA-Z_]/) ? true : false;
                 if (
                   paramIsSymbolName &&
-                  !this._isDebugMethod(newParameter) &&
-                  !this._isBinaryOperator(newParameter) &&
-                  !this._isUnaryOperator(newParameter) &&
-                  !this._isFloatConversion(newParameter) &&
-                  !this._isSpinBuiltinMethod(newParameter) &&
-                  !this._isSpinBuiltInVariable(newParameter)
+                  !this.parseUtils.isDebugMethod(newParameter) &&
+                  !this.parseUtils.isBinaryOperator(newParameter) &&
+                  !this.parseUtils.isUnaryOperator(newParameter) &&
+                  !this.parseUtils.isFloatConversion(newParameter) &&
+                  !this.parseUtils.isSpinBuiltinMethod(newParameter) &&
+                  !this.parseUtils.isSpinBuiltInVariable(newParameter)
                 ) {
                   this._logDEBUG("  -- rptDbg 3 unkParam=[" + newParameter + "]");
                   tokenSet.push({
@@ -3434,7 +3416,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
             isMethod = true;
           }
           const constantPart: string = possibleNameSet[1];
-          if (this._isStorageType(constantPart)) {
+          if (this.parseUtils.isStorageType(constantPart)) {
             // FIXME: UNDONE remove when syntax see this correctly
             const nameOffset: number = line.indexOf(constantPart, referenceOffset + refPart.length + 1);
             this._logSPIN("  --  rOr rhs storageType=[" + constantPart + "](" + (nameOffset + 1) + ")");
@@ -3473,7 +3455,7 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     let bNeedSingleQuoteProcessing: boolean = true;
     if (tokenStringSet.length > 0) {
       // see if we have sgl quites outside if dbl-quote strings
-      const nonStringLine: string = this._removeDoubleQuotedStrings(debugStatementStr);
+      const nonStringLine: string = this.parseUtils.removeDoubleQuotedStrings(debugStatementStr);
       bNeedSingleQuoteProcessing = nonStringLine.indexOf("'") != -1;
     }
     if (bNeedSingleQuoteProcessing) {
@@ -3575,22 +3557,6 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     return tokenSet;
   }
 
-  private _getSingleQuotedString(currentOffset: number, searchText: string): string {
-    let nextString: string = "";
-    const stringStartOffset: number = searchText.indexOf("'", currentOffset);
-    if (stringStartOffset != -1) {
-      this._logDEBUG("  -- _getSingleQuotedString(" + currentOffset + ", [" + searchText + "])");
-      const stringEndOffset: number = searchText.indexOf("'", stringStartOffset + 1);
-      if (stringEndOffset != -1) {
-        nextString = searchText.substring(stringStartOffset, stringEndOffset + 1);
-      }
-    }
-    if (nextString.length > 0) {
-      this._logDEBUG("  -- debug() gsqs nextString=[" + nextString + "](" + nextString.length + ")");
-    }
-    return nextString;
-  }
-
   private _reportDebugDblQuoteStrings(lineNumber: number, line: string, debugStatementStr: string): IParsedToken[] {
     const tokenSet: IParsedToken[] = [];
     // find all strings in debug() statement but for now just do first...
@@ -3676,6 +3642,22 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     }
     if (nextString.length > 0) {
       this._logDEBUG("  -- debug() gdqs nextString=[" + nextString + "](" + nextString.length + ")");
+    }
+    return nextString;
+  }
+
+  private _getSingleQuotedString(currentOffset: number, searchText: string): string {
+    let nextString: string = "";
+    const stringStartOffset: number = searchText.indexOf("'", currentOffset);
+    if (stringStartOffset != -1) {
+      this._logDEBUG("  -- gsqs(" + currentOffset + ", [" + searchText + "])");
+      const stringEndOffset: number = searchText.indexOf("'", stringStartOffset + 1);
+      if (stringEndOffset != -1) {
+        nextString = searchText.substring(stringStartOffset, stringEndOffset + 1);
+      }
+    }
+    if (nextString.length > 0) {
+      this._logDEBUG("  -- debug() gsqs nextString=[" + nextString + "](" + nextString.length + ")");
     }
     return nextString;
   }
@@ -3831,106 +3813,11 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     };
   }
 
-  private _getNonInlineCommentLine(line: string): string {
-    // NEW remove {comment} and {{comment}} single-line elements too
-    let nonInlineCommentStr: string = line;
-    // TODO: UNDONE make this into loop to find all single line {} or {{}} comments
-    const startDoubleBraceOffset: number = nonInlineCommentStr.indexOf("{{");
-    if (startDoubleBraceOffset != -1) {
-      const endDoubleBraceOffset: number = nonInlineCommentStr.indexOf("}}", startDoubleBraceOffset + 2);
-      if (endDoubleBraceOffset != -1) {
-        // remove this comment
-        const badElement = nonInlineCommentStr.substr(startDoubleBraceOffset, endDoubleBraceOffset - startDoubleBraceOffset + 1);
-        //this._logMessage('  -- badElement=[' + badElement + ']');
-        nonInlineCommentStr = nonInlineCommentStr.replace(badElement, " ".repeat(badElement.length));
-      }
-    }
-    const startSingleBraceOffset: number = nonInlineCommentStr.indexOf("{");
-    if (startSingleBraceOffset != -1) {
-      const endSingleBraceOffset: number = nonInlineCommentStr.indexOf("}", startSingleBraceOffset + 1);
-      if (endSingleBraceOffset != -1) {
-        // remove this comment
-        const badElement = nonInlineCommentStr.substr(startSingleBraceOffset, endSingleBraceOffset - startSingleBraceOffset + 1);
-        //this._logMessage('  -- badElement=[' + badElement + ']');
-        nonInlineCommentStr = nonInlineCommentStr.replace(badElement, " ".repeat(badElement.length));
-      }
-    }
-    //if (nonInlineCommentStr.length != line.length) {
-    //    this._logMessage('  -- NIC line [' + line + ']');
-    //    this._logMessage('  --          [' + nonInlineCommentStr + ']');
-    //}
-    return nonInlineCommentStr;
-  }
-
-  private _getNonCommentLineReturnComment(lineNumber: number, startingOffset: number, line: string, tokenSet: IParsedToken[]): string {
-    // skip Past Whitespace
-    let currentOffset: number = this._skipWhite(line, startingOffset);
-    const nonCommentLHSStr = this._getNonCommentLineRemainder(currentOffset, line);
-    // now record the comment if we have one
-    const commentRHSStrOffset: number = currentOffset + nonCommentLHSStr.length;
-    const commentOffset: number = line.indexOf("'", commentRHSStrOffset);
-    const bHaveDocComment: boolean = line.indexOf("''", commentOffset) != -1;
-    if (commentOffset != -1 && !bHaveDocComment) {
-      const newToken: IParsedToken = {
-        line: lineNumber,
-        startCharacter: commentOffset,
-        length: line.length - commentOffset + 1,
-        ptTokenType: "comment",
-        ptTokenModifiers: ["line"],
-      };
-      tokenSet.push(newToken);
-    }
-    return nonCommentLHSStr;
-  }
-
-  private _getNonCommentLineRemainder(startingOffset: number, line: string): string {
-    let nonCommentRHSStr: string = line;
-    //this._logMessage('  -- gnclr ofs=' + startingOffset + '[' + line + '](' + line.length + ')');
-    // TODO: UNDONE make this into loop to find first ' not in string
-    if (line.length - startingOffset > 0) {
-      //this._logMessage('- gnclr startingOffset=[' + startingOffset + '], startingOffset=[' + line + ']');
-      let currentOffset: number = this._skipWhite(line, startingOffset);
-      // get line parts - we only care about first one
-      let beginCommentOffset: number = line.indexOf("'", currentOffset);
-      if (beginCommentOffset != -1) {
-        // have single quote, is it within quoted string?
-        const startDoubleQuoteOffset: number = line.indexOf('"', currentOffset);
-        if (startDoubleQuoteOffset != -1) {
-          const nonStringLine: string = this._removeDoubleQuotedStrings(line, false); // false disabled debug output
-          beginCommentOffset = nonStringLine.indexOf("'", currentOffset);
-        }
-      }
-      if (beginCommentOffset === -1) {
-        beginCommentOffset = line.indexOf("{", currentOffset);
-      }
-      const nonCommentEOL: number = beginCommentOffset != -1 ? beginCommentOffset - 1 : line.length - 1;
-      //this._logMessage('- gnclr startingOffset=[' + startingOffset + '], currentOffset=[' + currentOffset + ']');
-      nonCommentRHSStr = line.substr(currentOffset, nonCommentEOL - currentOffset + 1).trim();
-      //this._logMessage('- gnclr nonCommentRHSStr=[' + startingOffset + ']');
-
-      const singleLineMultiBeginOffset: number = nonCommentRHSStr.indexOf("{", currentOffset);
-      if (singleLineMultiBeginOffset != -1) {
-        const singleLineMultiEndOffset: number = nonCommentRHSStr.indexOf("}", singleLineMultiBeginOffset);
-        if (singleLineMultiEndOffset != -1) {
-          const oneLineMultiComment: string = nonCommentRHSStr.substr(singleLineMultiBeginOffset, singleLineMultiEndOffset - singleLineMultiBeginOffset + 1);
-          nonCommentRHSStr = nonCommentRHSStr.replace(oneLineMultiComment, "").trim();
-        }
-      }
-    } else if (line.length - startingOffset == 0) {
-      nonCommentRHSStr = "";
-    }
-    //if (line.substr(startingOffset).length != nonCommentRHSStr.length) {
-    //    this._logMessage('  -- NCLR line [' + line.substr(startingOffset) + ']');
-    //    this._logMessage('  --           [' + nonCommentRHSStr + ']');
-    //}
-    return nonCommentRHSStr;
-  }
-
   private _getDebugStatement(startingOffset: number, line: string): string {
-    let currentOffset: number = this._skipWhite(line, startingOffset);
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
     let debugNonCommentStr: string = line;
     let openParenOffset: number = line.indexOf("(", currentOffset);
-    let closeParenOffset: number = this._indexOfMatchingCloseParen(line, openParenOffset);
+    let closeParenOffset: number = this.parseUtils.indexOfMatchingCloseParen(line, openParenOffset);
     if (line.length - startingOffset > 0 && openParenOffset != -1 && closeParenOffset != -1) {
       // have scope of debug line - remove trailing comment, trim it and return it
       let commentOffset: number = line.indexOf("'", closeParenOffset + 1);
@@ -3952,84 +3839,9 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     return debugNonCommentStr;
   }
 
-  private _indexOfMatchingCloseParen(line: string, openParenOffset: number): number {
-    let desiredCloseOffset: number = -1;
-    let nestingDepth: number = 1;
-    for (let offset = openParenOffset + 1; offset < line.length; offset++) {
-      if (line.substring(offset, offset + 1) == "(") {
-        nestingDepth++;
-      } else if (line.substring(offset, offset + 1) == ")") {
-        nestingDepth--;
-        if (nestingDepth == 0) {
-          // we closed the inital open
-          desiredCloseOffset = offset;
-          break; // done, get outta here
-        }
-      }
-    }
-    // this._logMessage('  -- iomcp line=[' + line + ']');
-    // this._logMessage('  --       open=(' + openParenOffset + '), close=(' + desiredCloseOffset + ')');
-    return desiredCloseOffset;
-  }
-
-  private _getNonDocCommentLineRemainder(startingOffset: number, line: string): string {
-    let nonDocCommentRHSStr: string = line;
-    //this._logMessage('  -- gnclr ofs=' + startingOffset + '[' + line + '](' + line.length + ')');
-    // TODO: UNDONE make this into loop to find first ' not in string
-    if (line.length - startingOffset > 0) {
-      const nonCommentEOL: number = line.length - 1;
-      //this._logMessage('- gnclr startingOffset=[' + startingOffset + '], currentOffset=[' + currentOffset + ']');
-      nonDocCommentRHSStr = line.substr(startingOffset, nonCommentEOL - startingOffset + 1).trim();
-      //this._logMessage('- gnclr nonCommentRHSStr=[' + startingOffset + ']');
-
-      const singleLineMultiBeginOffset: number = nonDocCommentRHSStr.indexOf("{", startingOffset);
-      if (singleLineMultiBeginOffset != -1) {
-        const singleLineMultiEndOffset: number = nonDocCommentRHSStr.indexOf("}", singleLineMultiBeginOffset);
-        if (singleLineMultiEndOffset != -1) {
-          const oneLineMultiComment: string = nonDocCommentRHSStr.substr(singleLineMultiBeginOffset, singleLineMultiEndOffset - singleLineMultiBeginOffset + 1);
-          nonDocCommentRHSStr = nonDocCommentRHSStr.replace(oneLineMultiComment, "").trim();
-        }
-      }
-    } else if (line.length - startingOffset == 0) {
-      nonDocCommentRHSStr = "";
-    }
-    //if (line.substr(startingOffset).length != nonCommentRHSStr.length) {
-    //    this._logMessage('  -- NCLR line [' + line.substr(startingOffset) + ']');
-    //    this._logMessage('  --           [' + nonCommentRHSStr + ']');
-    //}
-    return nonDocCommentRHSStr;
-  }
-
-  private _getNonWhiteDataInitLineParts(line: string): string[] {
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\,\[\]\(\)\+\-\/\<\>\|\*\@]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getNonWhiteCONLineParts(line: string): string[] {
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^  \t\(\)\*\+\-\/\>\<\=]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getNonWhitePasmLineParts(line: string): string[] {
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
-    let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\,\(\)\[\]\<\>\=\?\!\^\+\*\&\|\-\\\#\@\/]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
   private _getNonWhiteSpinLineParts(line: string): IFilteredStrings {
     //                                     split(/[ \t\-\:\,\+\[\]\@\(\)\!\*\=\<\>\&\|\?\\\~\#\^\/]/);
-    const nonEqualsLine: string = this._removeDoubleQuotedStrings(line);
+    const nonEqualsLine: string = this.parseUtils.removeDoubleQuotedStrings(line);
     let lineParts: string[] | null = nonEqualsLine.match(/[^ \t\-\:\,\+\[\]\@\(\)\!\*\=\<\>\&\|\?\\\~\#\^\/]+/g);
     if (lineParts === null) {
       lineParts = [];
@@ -4040,163 +3852,25 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     };
   }
 
-  private _removeDoubleQuotedStrings(line: string, showDebug: boolean = true): string {
-    //this._logMessage('- RQS line [' + line + ']');
-    let trimmedLine: string = line;
-    //this._logMessage('- RQS line [' + line + ']');
-    const doubleQuote: string = '"';
-    let quoteStartOffset: number = 0; // value doesn't matter
-    let didRemove: boolean = false;
-    while ((quoteStartOffset = trimmedLine.indexOf(doubleQuote)) != -1) {
-      const quoteEndOffset: number = trimmedLine.indexOf(doubleQuote, quoteStartOffset + 1);
-      //this._logMessage('  -- quoteStartOffset=[' + quoteStartOffset + '] quoteEndOffset=[' + quoteEndOffset + ']');
-      if (quoteEndOffset != -1) {
-        const badElement = trimmedLine.substr(quoteStartOffset, quoteEndOffset - quoteStartOffset + 1);
-        //this._logMessage('  -- badElement=[' + badElement + ']');
-        trimmedLine = trimmedLine.replace(badElement, "#".repeat(badElement.length));
-        didRemove = showDebug ? true : false;
-        //this._logMessage('-         post[' + trimmedLine + ']');
-      } else {
-        break; // we don't handle a single double-quote
-      }
+  private _getNonCommentLineReturnComment(lineNumber: number, startingOffset: number, line: string, tokenSet: IParsedToken[]): string {
+    // skip Past Whitespace
+    let currentOffset: number = this.parseUtils.skipWhite(line, startingOffset);
+    const nonCommentLHSStr = this.parseUtils.getNonCommentLineRemainder(currentOffset, line);
+    // now record the comment if we have one
+    const commentRHSStrOffset: number = currentOffset + nonCommentLHSStr.length;
+    const commentOffset: number = line.indexOf("'", commentRHSStrOffset);
+    const bHaveDocComment: boolean = line.indexOf("''", commentOffset) != -1;
+    if (commentOffset != -1 && !bHaveDocComment) {
+      const newToken: IParsedToken = {
+        line: lineNumber,
+        startCharacter: commentOffset,
+        length: line.length - commentOffset + 1,
+        ptTokenType: "comment",
+        ptTokenModifiers: ["line"],
+      };
+      tokenSet.push(newToken);
     }
-
-    //if (didRemove) {
-    //    this._logMessage('  -- RQS line [' + line + ']');
-    //    this._logMessage('  --          [' + trimmedLine + ']');
-    //}
-
-    return trimmedLine;
-  }
-
-  private _getNonWhiteLineParts(line: string): string[] {
-    let lineParts: string[] | null = line.match(/[^ \t]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _removeDebugSingleQuotedStrings(line: string, showDebug: boolean = true): string {
-    // remove single-quoted strings from keyword processing
-    //  Ex #1:   ' a string '
-    //  Ex #2:   ' a string up to [`(var)]
-    //  Ex #3:   [)] a string after var'
-    //  Ex #4:   [)] a string up to another [`(var)]
-    //  Ex #5:   debug(`scope_xy xy size 200 range 1000 samples 200 dotsize 5 'Goertzel' `dly(#200))
-    this._logMessage("- RQS line [" + line + "]");
-    let trimmedLine: string = line;
-    this._logMessage("  -- trim line [" + trimmedLine + "]");
-    const chrSingleQuote: string = "'";
-    const chrBackTic: string = "`";
-    const chrOpenParen: string = "(";
-    const chrCloseParen: string = ")";
-    let didRemove: boolean = false;
-    const firstOpenParenOffset: number = trimmedLine.indexOf(chrOpenParen, 0);
-    // skip past tic-open pairs and their closes
-    let nextBackTic: number = trimmedLine.indexOf(chrBackTic, 0);
-    let secondsBackTic: number = trimmedLine.indexOf(chrBackTic, nextBackTic + 1);
-    let lastCloseParenOffset: number = trimmedLine.indexOf(chrCloseParen, 0);
-    // this._logMessage('  -- 1 nextBackTic=[' + nextBackTic + '] lastCloseParenOffset=[' + lastCloseParenOffset + ']');
-    while (nextBackTic != -1) {
-      // if we have another back-tic before any parens skip the first one it's only a debug term id marker
-      if (secondsBackTic < lastCloseParenOffset) {
-        nextBackTic = secondsBackTic;
-        // this._logMessage('  -- 1b nextBackTic=[' + nextBackTic + '] lastCloseParenOffset=[' + lastCloseParenOffset + ']');
-      }
-      // if we have `( followed by ) then skip this close, look for next
-      if (lastCloseParenOffset > nextBackTic) {
-        // look for next close
-        // this._logMessage('  -- SKIP backticOpenOffset=[' + nextBackTic + '] lastCloseParenOffset=[' + lastCloseParenOffset + ']');
-        lastCloseParenOffset = trimmedLine.indexOf(chrCloseParen, lastCloseParenOffset + 1);
-      }
-      nextBackTic = trimmedLine.indexOf(chrBackTic, nextBackTic + 1);
-    }
-    // by now lastCloseParenOffset should point to end of statement within line
-    let quoteStartOffset: number = trimmedLine.indexOf(chrSingleQuote, 0);
-    // this._logMessage('  -- 2 quoteStartOffset=[' + quoteStartOffset + '] lastCloseParenOffset=[' + lastCloseParenOffset + ']');
-    while (quoteStartOffset != -1) {
-      let bHaveBackTic: boolean = false;
-      const quoteEndOffset: number = trimmedLine.indexOf(chrSingleQuote, quoteStartOffset + 1);
-      if (quoteEndOffset > lastCloseParenOffset) {
-        break; // nothing beyond end of line please
-      }
-      this._logMessage("  -- quoteStartOffset=[" + quoteStartOffset + "] quoteEndOffset=[" + quoteEndOffset + "]");
-      if (quoteEndOffset != -1) {
-        // any more strings? on this line?
-        let badElement: string = trimmedLine.substr(quoteStartOffset, quoteEndOffset - quoteStartOffset + 1);
-        let backTicOffset: number = trimmedLine.indexOf(chrBackTic, quoteStartOffset);
-        this._logMessage("  -- RdsQS backTicOffset=[" + backTicOffset + "], quoteEndOffset=[" + quoteEndOffset + "], badElement=[" + badElement + "]");
-        if (backTicOffset != -1 && backTicOffset < quoteEndOffset) {
-          bHaveBackTic = true;
-          badElement = trimmedLine.substr(quoteStartOffset, backTicOffset - quoteStartOffset);
-          if (badElement.length > 0) {
-            // badElement = badElement.replace(chrBackTic, '');    // remove bacTicks
-          }
-          // this._logMessage('  -- RdsQS 2 backTicOffset=[' + backTicOffset + '], quoteEndOffset=[' + quoteEndOffset + '], badElement=[' + badElement + ']');
-        }
-        this._logMessage("  -- RdsQS badElement=[" + badElement + "]");
-        trimmedLine = trimmedLine.replace(badElement, "#".repeat(badElement.length));
-        didRemove = showDebug ? true : false;
-        this._logMessage("  -- RdsQS post[" + trimmedLine + "]");
-        // handle  #3 and #4 cases
-        if (bHaveBackTic) {
-          const closeParenOffset: number = trimmedLine.indexOf(chrCloseParen, backTicOffset + 1);
-          // have case #2?
-          backTicOffset = trimmedLine.indexOf(chrBackTic, closeParenOffset);
-          if (backTicOffset != -1) {
-            // we have another backtic, just return to top of loop
-            quoteStartOffset = closeParenOffset + 1;
-          } else if (closeParenOffset != -1) {
-            // let's skip to triling string
-            quoteStartOffset = closeParenOffset + 1;
-            if (quoteStartOffset < quoteEndOffset) {
-              badElement = trimmedLine.substr(quoteStartOffset, quoteEndOffset - quoteStartOffset + 1);
-              //this._logMessage('  -- RdsQS rhs quoteStartOffset=[' + quoteStartOffset + '], quoteEndOffset=[' + quoteEndOffset + '], badElement=[' + badElement + ']');
-              trimmedLine = trimmedLine.replace(badElement, "#".repeat(badElement.length));
-              if (showDebug) {
-                didRemove = true;
-              }
-              //this._logMessage('  -- RdsQS rhs post[' + trimmedLine + ']');
-            }
-            // finished this quote pair, find start of next possible pair
-            quoteStartOffset = trimmedLine.indexOf(chrSingleQuote, quoteEndOffset + 1);
-          }
-        }
-      } else {
-        break; // we don't handle a single double-quote
-      }
-    }
-
-    if (didRemove) {
-      this._logMessage("  -- RdsQS line [" + line + "]");
-      this._logMessage("  --            [" + trimmedLine + "]");
-    }
-
-    return trimmedLine;
-  }
-
-  private _getDebugNonWhiteLineParts(line: string): string[] {
-    // remove douple and then any single quotes string from display list
-    //this._logMessage('  -- gdnwlp raw-line [' + line + ']');
-    const nonDblStringLine: string = this._removeDoubleQuotedStrings(line);
-    this._logMessage("  -- gdnwlp nonDblStringLine=[" + nonDblStringLine + "]");
-    const nonSglStringLine: string = this._removeDebugSingleQuotedStrings(nonDblStringLine, false);
-    this._logMessage("  -- gdnwlp nonSglStringLine=[" + nonSglStringLine + "]");
-    let lineParts: string[] | null = nonSglStringLine.match(/[^ ,@\[\]\+\-\*\/\<\>\t\(\)\!\?\~]+/g);
-    //let lineParts: string[] | null = line.match(/[^ ,@\[\]\+\-\*\/\<\>\t\(\)]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
-  }
-
-  private _getCommaDelimitedNonWhiteLineParts(line: string): string[] {
-    let lineParts: string[] | null = line.match(/[^ \t,]+/g);
-    if (lineParts === null) {
-      lineParts = [];
-    }
-    return lineParts;
+    return nonCommentLHSStr;
   }
 
   // -------------------------------------------------------------------------
@@ -4376,1553 +4050,8 @@ export class Spin2DocumentSemanticTokensProvider implements vscode.DocumentSeman
     return updatedModifiers;
   }
 
-  // -------------------------------------------------------------------------
-  // keyword checks
-  private _isBuiltinReservedWord(name: string): boolean {
-    // streamer constants, smart-pin constants
-    const builtinNamesOfNote: string[] = [
-      // streamer names
-      "x_16p_2dac8_wfword",
-      "x_16p_4dac4_wfword",
-      "x_1adc8_0p_1dac8_wfbyte",
-      "x_1adc8_8p_2dac8_wfword",
-      "x_1p_1dac1_wfbyte",
-      "x_2adc8_0p_2dac8_wfword",
-      "x_2adc8_16p_4dac8_wflong",
-      "x_2p_1dac2_wfbyte",
-      "x_2p_2dac1_wfbyte",
-      "x_32p_4dac8_wflong",
-      "x_4adc8_0p_4dac8_wflong",
-      "x_4p_1dac4_wfbyte",
-      "x_4p_2dac2_wfbyte",
-      "x_4p_4dac1_wfbyte",
-      "x_8p_1dac8_wfbyte",
-      "x_8p_2dac4_wfbyte",
-      "x_8p_4dac2_wfbyte",
-      "x_alt_off",
-      "x_alt_on",
-      "x_dacs_0n0_0n0",
-      "x_dacs_0n0_x_x",
-      "x_dacs_0_0_0_0",
-      "x_dacs_0_0_x_x",
-      "x_dacs_0_x_x_x",
-      "x_dacs_1n1_0n0",
-      "x_dacs_1_0_1_0",
-      "x_dacs_1_0_x_x",
-      "x_dacs_3_2_1_0",
-      "x_dacs_off",
-      "x_dacs_x_0_x_x",
-      "x_dacs_x_x_0n0",
-      "x_dacs_x_x_0_0",
-      "x_dacs_x_x_0_x",
-      "x_dacs_x_x_1_0",
-      "x_dacs_x_x_x_0",
-      "x_dds_goertzel_sinc1",
-      "x_dds_goertzel_sinc2",
-      "x_imm_16x2_1dac2",
-      "x_imm_16x2_2dac1",
-      "x_imm_16x2_lut",
-      "x_imm_1x32_4dac8",
-      "x_imm_2x16_2dac8",
-      "x_imm_2x16_4dac4",
-      "x_imm_32x1_1dac1",
-      "x_imm_32x1_lut",
-      "x_imm_4x8_1dac8",
-      "x_imm_4x8_2dac4",
-      "x_imm_4x8_4dac2",
-      "x_imm_4x8_lut",
-      "x_imm_8x4_1dac4",
-      "x_imm_8x4_2dac2",
-      "x_imm_8x4_4dac1",
-      "x_imm_8x4_lut",
-      "x_pins_off",
-      "x_pins_on",
-      "x_rfbyte_1p_1dac1",
-      "x_rfbyte_2p_1dac2",
-      "x_rfbyte_2p_2dac1",
-      "x_rfbyte_4p_1dac4",
-      "x_rfbyte_4p_2dac2",
-      "x_rfbyte_4p_4dac1",
-      "x_rfbyte_8p_1dac8",
-      "x_rfbyte_8p_2dac4",
-      "x_rfbyte_8p_4dac2",
-      "x_rfbyte_luma8",
-      "x_rfbyte_rgb8",
-      "x_rfbyte_rgbi8",
-      "x_rflong_16x2_lut",
-      "x_rflong_32p_4dac8",
-      "x_rflong_32x1_lut",
-      "x_rflong_4x8_lut",
-      "x_rflong_8x4_lut",
-      "x_rflong_rgb24",
-      "x_rfword_16p_2dac8",
-      "x_rfword_16p_4dac4",
-      "x_rfword_rgb16",
-      "x_write_off",
-      "x_write_on",
-      // smart pin names
-      "p_adc",
-      "p_adc_100x",
-      "p_adc_10x",
-      "p_adc_1x",
-      "p_adc_30x",
-      "p_adc_3x",
-      "p_adc_ext",
-      "p_adc_float",
-      "p_adc_gio",
-      "p_adc_scope",
-      "p_adc_vio",
-      "p_async_io",
-      "p_async_rx",
-      "p_async_tx",
-      "p_bitdac",
-      "p_channel",
-      "p_compare_ab",
-      "p_compare_ab_fb",
-      "p_counter_highs",
-      "p_counter_periods",
-      "p_counter_ticks",
-      "p_count_highs",
-      "p_count_rises",
-      "p_dac_124r_3v",
-      "p_dac_600r_2v",
-      "p_dac_75r_2v",
-      "p_dac_990r_3v",
-      "p_dac_dither_pwm",
-      "p_dac_dither_rnd",
-      "p_dac_noise",
-      "p_events_ticks",
-      "p_high_100ua",
-      "p_high_10ua",
-      "p_high_150k",
-      "p_high_15k",
-      "p_high_1k5",
-      "p_high_1ma",
-      "p_high_fast",
-      "p_high_float",
-      "p_high_ticks",
-      "p_invert_a",
-      "p_invert_b",
-      "p_invert_in",
-      "p_invert_output",
-      "p_level_a",
-      "p_level_a_fbn",
-      "p_level_a_fbp",
-      "p_local_a",
-      "p_local_b",
-      "p_logic_a",
-      "p_logic_a_fb",
-      "p_logic_b_fb",
-      "p_low_100ua",
-      "p_low_10ua",
-      "p_low_150k",
-      "p_low_15k",
-      "p_low_1k5",
-      "p_low_1ma",
-      "p_low_fast",
-      "p_low_float",
-      "p_minus1_a",
-      "p_minus1_b",
-      "p_minus2_a",
-      "p_minus2_b",
-      "p_minus3_a",
-      "p_minus3_b",
-      "p_nco_duty",
-      "p_nco_freq",
-      "p_normal",
-      "p_oe",
-      "p_outbit_a",
-      "p_outbit_b",
-      "p_periods_highs",
-      "p_periods_ticks",
-      "p_plus1_a",
-      "p_plus1_b",
-      "p_plus2_a",
-      "p_plus2_b",
-      "p_plus3_a",
-      "p_plus3_b",
-      "p_pulse",
-      "p_pwm_sawtooth",
-      "p_pwm_smps",
-      "p_pwm_triangle",
-      "p_quadrature",
-      "p_reg_down",
-      "p_reg_up",
-      "p_repository",
-      "p_schmitt_a",
-      "p_schmitt_a_fb",
-      "p_schmitt_b_fb",
-      "p_state_ticks",
-      "p_sync_io",
-      "p_sync_rx",
-      "p_sync_tx",
-      "p_transition",
-      "p_true_a",
-      "p_true_b",
-      "p_true_in",
-      "p_true_output",
-      "p_tt_00",
-      "p_tt_01",
-      "p_tt_10",
-      "p_tt_11",
-      "p_usb_pair",
-      // event names
-      "event_atn",
-      "event_ct1",
-      "event_ct2",
-      "event_ct3",
-      "event_fbw",
-      "event_int",
-      "event_pat",
-      "event_qmt",
-      "event_se1",
-      "event_se2",
-      "event_se3",
-      "event_se4",
-      "event_xfi",
-      "event_xmt",
-      "event_xrl",
-      "event_xro",
-      //
-      "pr0",
-      "pr1",
-      "pr2",
-      "pr3",
-      "pr4",
-      "pr5",
-      "pr6",
-      "pr7",
-      "ijmp1",
-      "ijmp2",
-      "ijmp3",
-      "iret1",
-      "iret2",
-      "iret3",
-      "pa",
-      "pb",
-      "ptra",
-      "ptrb",
-      "dira",
-      "dirb",
-      "outa",
-      "outb",
-      "ina",
-      "inb",
-    ];
-    const reservedStatus: boolean = builtinNamesOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isDebugMethod(name: string): boolean {
-    const debugMethodOfNote: string[] = [
-      "zstr",
-      "lstr",
-      "udec",
-      "udec_byte",
-      "udec_word",
-      "udec_long",
-      "udec_reg_array",
-      "udec_byte_array",
-      "udec_word_array",
-      "udec_long_array",
-      "sdec",
-      "sdec_byte",
-      "sdec_word",
-      "sdec_long",
-      "sdec_reg_array",
-      "sdec_byte_array",
-      "sdec_word_array",
-      "sdec_long_array",
-      "uhex",
-      "uhex_byte",
-      "uhex_word",
-      "uhex_long",
-      "uhex_reg_array",
-      "uhex_byte_array",
-      "uhex_word_array",
-      "uhex_long_array",
-      "shex",
-      "shex_byte",
-      "shex_word",
-      "shex_long",
-      "shex_reg_array",
-      "shex_byte_array",
-      "shex_word_array",
-      "shex_long_array",
-      "ubin",
-      "ubin_byte",
-      "ubin_word",
-      "ubin_long",
-      "ubin_reg_array",
-      "ubin_byte_array",
-      "ubin_word_array",
-      "ubin_long_array",
-      "sbin",
-      "sbin_byte",
-      "sbin_word",
-      "sbin_long",
-      "sbin_reg_array",
-      "sbin_byte_array",
-      "sbin_word_array",
-      "sbin_long_array",
-      "fdec",
-      "fdec_array",
-      "fdec_reg_array",
-      "dly",
-      "pc_key",
-      "pc_mouse",
-      "if",
-      "ifnot",
-    ];
-    const searchName: string = name.endsWith("_") ? name.substr(0, name.length - 1) : name;
-    const reservedStatus: boolean = debugMethodOfNote.indexOf(searchName.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isDebugInvocation(name: string): boolean {
-    const debugExec: string[] = [
-      // debug overridable CONSTANTS
-      "debug_main",
-      "debug_coginit",
-      "debug",
-    ];
-    const reservedStatus: boolean = debugExec.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isDebugSymbol(name: string): boolean {
-    const debugSymbolOfNote: string[] = [
-      // debug overridable CONSTANTS
-      "download_baud",
-      "debug_cogs",
-      "debug_delay",
-      "debug_pin_tx",
-      "debug_pin_rx",
-      "debug_baud",
-      "debug_timestamp",
-      "debug_log_size",
-      "debug_left",
-      "debug_top",
-      "debug_width",
-      "debug_height",
-      "debug_display_left",
-      "debug_display_top",
-      "debug_windows_off",
-    ];
-    const searchName: string = name.endsWith("_") ? name.substr(0, name.length - 1) : name;
-    const reservedStatus: boolean = debugSymbolOfNote.indexOf(searchName.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinBuiltInVariable(name: string): boolean {
-    const spinVariablesOfNote: string[] = [
-      "clkmode",
-      "clkfreq",
-      "varbase",
-      "pr0",
-      "pr1",
-      "pr2",
-      "pr3",
-      "pr4",
-      "pr5",
-      "pr6",
-      "pr7",
-      "ijmp1",
-      "ijmp2",
-      "ijmp3",
-      "iret1",
-      "iret2",
-      "iret3",
-      "pa",
-      "pb",
-      "ptra",
-      "ptrb",
-      "dira",
-      "dirb",
-      "outa",
-      "outb",
-      "ina",
-      "inb",
-    ];
-    let reservedStatus: boolean = spinVariablesOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinReservedWord(name: string): boolean {
-    const spinInstructionsOfNote: string[] = [
-      "reg",
-      "float",
-      "round",
-      "trunc",
-      "nan",
-      "clkmode",
-      "clkfreq",
-      "varbase",
-      "clkmode_",
-      "clkfreq_",
-      "if",
-      "ifnot",
-      "elseif",
-      "elseifnot",
-      "else",
-      "while",
-      "repeat",
-      "until",
-      "from",
-      "to",
-      "step",
-      "next",
-      "quit",
-      "case",
-      "case_fast",
-      "other",
-      "abort",
-      "return",
-      "true",
-      "false",
-      "posx",
-      "negx",
-      "pi",
-    ];
-    let reservedStatus: boolean = spinInstructionsOfNote.indexOf(name.toLowerCase()) != -1;
-    if (reservedStatus == false) {
-      reservedStatus = this._isBinaryOperator(name);
-    }
-    if (reservedStatus == false) {
-      reservedStatus = this._isUnaryOperator(name);
-    }
-    return reservedStatus;
-  }
-  private _isCoginitReservedSymbol(name: string): boolean {
-    const coginitSymbolOfNote: string[] = [
-      //
-      "newcog",
-      "cogexec",
-      "hubexec",
-      "cogexec_new",
-      "hubexec_new",
-      "cogexec_new_pair",
-      "hubexec_new_pair",
-    ];
-    const reservedStatus: boolean = coginitSymbolOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isBinaryOperator(name: string): boolean {
-    const binaryOperationsOfNote: string[] = ["sar", "ror", "rol", "rev", "zerox", "signx", "sca", "scas", "frac", "addbits", "addpins", "and", "or", "xor"];
-    const reservedStatus: boolean = binaryOperationsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isUnaryOperator(name: string): boolean {
-    const unaryOperationsOfNote: string[] = ["not", "abs", "fabs", "encod", "decod", "bmask", "ones", "sqrt", "fsqrt", "qlog", "qexp"];
-    const reservedStatus: boolean = unaryOperationsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isFloatConversion(name: string): boolean {
-    const floatConversionOfNote: string[] = ["float", "round", "trunc"];
-    const reservedStatus: boolean = floatConversionOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isPasmModczOperand(name: string): boolean {
-    const pasmModczOperand: string[] = [
-      "_clr",
-      "_nc_and_nz",
-      "_nz_and_nc",
-      " _gt",
-      "_nc_and_z",
-      "_z_and_nc",
-      "_nc",
-      "_ge",
-      "_c_and_nz",
-      "_nz_and_c",
-      "_nz",
-      "_ne",
-      "_c_ne_z",
-      "_z_ne_c",
-      "_nc_or_nz",
-      "_nz_or_nc",
-      "_c_and_z",
-      "_z_and_c",
-      "_c_eq_z",
-      "_z_eq_c",
-      "_z",
-      "_e",
-      "_nc_or_z",
-      "_z_or_nc",
-      "_c",
-      "_lt",
-      "_c_or_nz",
-      "_nz_or_c",
-      "_c_or_z",
-      "_z_or_c",
-      "_le",
-      "_set",
-    ];
-    const reservedStatus: boolean = pasmModczOperand.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isSpinBuiltinMethod(name: string): boolean {
-    const spinMethodNames: string[] = [
-      "akpin",
-      "bytefill",
-      "bytemove",
-      "call",
-      "clkset",
-      "cogatn",
-      "cogchk",
-      "cogid",
-      "coginit",
-      "cogspin",
-      "cogstop",
-      "getct",
-      "getcrc",
-      "getregs",
-      "getrnd",
-      "getsec",
-      "getms",
-      "hubset",
-      "lockchk",
-      "locknew",
-      "lockrel",
-      "lockret",
-      "locktry",
-      "longfill",
-      "longmove",
-      "lookdown",
-      "lookdownz",
-      "lookup",
-      "lookupz",
-      "muldiv64",
-      "pinclear",
-      "pinf",
-      "pinfloat",
-      "pinh",
-      "pinhigh",
-      "pinl",
-      "pinlow",
-      "pinr",
-      "pinread",
-      "pinstart",
-      "pint",
-      "pintoggle",
-      "pinw",
-      "pinwrite",
-      "pollatn",
-      "pollct",
-      "polxy",
-      "rdpin",
-      "recv",
-      "regexec",
-      "regload",
-      "rotxy",
-      "rqpin",
-      "send",
-      "setregs",
-      "strcomp",
-      "strcopy",
-      "string",
-      "strsize",
-      "waitatn",
-      "waitct",
-      "waitms",
-      "waitus",
-      "wordfill",
-      "wordmove",
-      "wrpin",
-      "wxpin",
-      "wypin",
-      "xypol",
-      "qsin",
-      "qcos",
-    ];
-    const reservedStatus: boolean = spinMethodNames.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isFlexspinPreprocessorDirective(name: string): boolean {
-    const flexspinDirectiveOfNote: string[] = ["#define", "#ifdef", "#ifndef", "#else", "#elseifdef", "#elseifndef", "#endif", "#error", "#include", "#warn", "#undef"];
-    const reservedStatus: boolean = flexspinDirectiveOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isFlexspinReservedWord(name: string): boolean {
-    const flexspinReservedswordsOfNote: string[] = [
-      "__propeller__",
-      "__propeller2__",
-      "__p2__",
-      "__flexspin__",
-      "__spincvt__",
-      "__spin2pasm__",
-      "__spin2cpp__",
-      "__have_fcache__",
-      "__cplusplus__",
-      "__date__",
-      "__file__",
-      "__line__",
-      "__time__",
-      "__version__",
-      "__debug__",
-      "__output_asm__",
-      "__output_bytecode__",
-      "__output_c__",
-      "__output_cpp__",
-    ];
-    const reservedStatus: boolean = flexspinReservedswordsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isReservedPasmSymbols(name: string): boolean {
-    const reservedPasmSymbolNames: string[] = ["org", "orgf", "orgh", "fit", "end"];
-    const reservedStatus: boolean = reservedPasmSymbolNames.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isPasmReservedWord(name: string): boolean {
-    const pasmReservedswordsOfNote: string[] = [
-      "ijmp1",
-      "ijmp2",
-      "ijmp3",
-      "iret1",
-      "iret2",
-      "iret3",
-      "ptra",
-      "ptrb",
-      "addpins",
-      "clkfreq_",
-      "pa",
-      "pb",
-      "clkfreq",
-      "_clkfreq",
-      "round",
-      "float",
-      "trunc",
-      "dira",
-      "dirb",
-      "ina",
-      "inb",
-      "outa",
-      "outb",
-      "fvar",
-      "fvars",
-      "addbits",
-      "true",
-      "false",
-    ];
-    const reservedStatus: boolean = pasmReservedswordsOfNote.indexOf(name.toLowerCase()) != -1;
-    return reservedStatus;
-  }
-
-  private _isPasmInstruction(name: string): boolean {
-    const pasmInstructions: string[] = [
-      "abs",
-      "add",
-      "addct1",
-      "addct2",
-      "addct3",
-      "addpix",
-      "adds",
-      "addsx",
-      "addx",
-      "akpin",
-      "allowi",
-      "altb",
-      "altd",
-      "altgb",
-      "altgn",
-      "altgw",
-      "alti",
-      "altr",
-      "alts",
-      "altsb",
-      "altsn",
-      "altsw",
-      "and",
-      "andn",
-      "augd",
-      "augs",
-      "bitc",
-      "bith",
-      "bitl",
-      "bitnc",
-      "bitnot",
-      "bitnz",
-      "bitrnd",
-      "bitz",
-      "blnpix",
-      "bmask",
-      "brk",
-      "call",
-      "calla",
-      "callb",
-      "calld",
-      "callpa",
-      "callpb",
-      "cmp",
-      "cmpm",
-      "cmpr",
-      "cmps",
-      "cmpsub",
-      "cmpsx",
-      "cmpx",
-      "cogatn",
-      "cogbrk",
-      "cogid",
-      "coginit",
-      "cogstop",
-      "crcbit",
-      "crcnib",
-      "decmod",
-      "decod",
-      "dirc",
-      "dirh",
-      "dirl",
-      "dirnc",
-      "dirnot",
-      "dirnz",
-      "dirrnd",
-      "dirz",
-      "djf",
-      "djnf",
-      "djnz",
-      "djz",
-      "drvc",
-      "drvh",
-      "drvl",
-      "drvnc",
-      "drvnot",
-      "drvnz",
-      "drvrnd",
-      "drvz",
-      "encod",
-      "execf",
-      "fblock",
-      "fge",
-      "fges",
-      "fle",
-      "fles",
-      "fltc",
-      "flth",
-      "fltl",
-      "fltnc",
-      "fltnot",
-      "fltnz",
-      "fltrnd",
-      "fltz",
-      "getbrk",
-      "getbyte",
-      "getbyte",
-      "getct",
-      "getnib",
-      "getptr",
-      "getqx",
-      "getqy",
-      "getrnd",
-      "getrnd",
-      "getscp",
-      "getword",
-      "getword",
-      "getxacc",
-      "hubset",
-      "ijnz",
-      "ijz",
-      "incmod",
-      "jatn",
-      "jct1",
-      "jct2",
-      "jct3",
-      "jfbw",
-      "jint",
-      "jmp",
-      "jmprel",
-      "jnatn",
-      "jnct1",
-      "jnct2",
-      "jnct3",
-      "jnfbw",
-      "jnint",
-      "jnpat",
-      "jnqmt",
-      "jnse1",
-      "jnse2",
-      "jnse3",
-      "jnse4",
-      "jnxfi",
-      "jnxmt",
-      "jnxrl",
-      "jnxro",
-      "jpat",
-      "jqmt",
-      "jse1",
-      "jse2",
-      "jse3",
-      "jse4",
-      "jxfi",
-      "jxmt",
-      "jxrl",
-      "jxro",
-      "loc",
-      "locknew",
-      "lockrel",
-      "lockret",
-      "locktry",
-      "mergeb",
-      "mergew",
-      "mixpix",
-      "modc",
-      "modcz",
-      "modz",
-      "mov",
-      "movbyts",
-      "mul",
-      "mulpix",
-      "muls",
-      "muxc",
-      "muxnc",
-      "muxnibs",
-      "muxnits",
-      "muxnz",
-      "muxq",
-      "muxz",
-      "neg",
-      "negc",
-      "negnc",
-      "negnz",
-      "negz",
-      "nixint1",
-      "nixint2",
-      "nixint3",
-      "nop",
-      "not",
-      "ones",
-      "or",
-      "outc",
-      "outh",
-      "outl",
-      "outnc",
-      "outnot",
-      "outnz",
-      "outrnd",
-      "outz",
-      "pollatn",
-      "pollct1",
-      "pollct2",
-      "pollct3",
-      "pollfbw",
-      "pollint",
-      "pollpat",
-      "pollqmt",
-      "pollse1",
-      "pollse2",
-      "pollse3",
-      "pollse4",
-      "pollxfi",
-      "pollxmt",
-      "pollxrl",
-      "pollxro",
-      "pop",
-      "popa",
-      "popb",
-      "push",
-      "pusha",
-      "pushb",
-      "qdiv",
-      "qexp",
-      "qfrac",
-      "qlog",
-      "qmul",
-      "qrotate",
-      "qsqrt",
-      "qvector",
-      "rcl",
-      "rcr",
-      "rczl",
-      "rczr",
-      "rdbyte",
-      "rdfast",
-      "rdlong",
-      "rdlut",
-      "rdpin",
-      "rdword",
-      "rep",
-      "resi0",
-      "resi1",
-      "resi2",
-      "resi3",
-      "ret",
-      "reta",
-      "retb",
-      "reti0",
-      "reti1",
-      "reti2",
-      "reti3",
-      "rev",
-      "rfbyte",
-      "rflong",
-      "rfvar",
-      "rfvars",
-      "rfword",
-      "rgbexp",
-      "rgbsqz",
-      "rol",
-      "rolbyte",
-      "rolbyte",
-      "rolnib",
-      "rolword",
-      "rolword",
-      "ror",
-      "rqpin",
-      "sal",
-      "sar",
-      "sca",
-      "scas",
-      "setbyte",
-      "setcfrq",
-      "setci",
-      "setcmod",
-      "setcq",
-      "setcy",
-      "setd",
-      "setdacs",
-      "setint1",
-      "setint2",
-      "setint3",
-      "setluts",
-      "setnib",
-      "setpat",
-      "setpiv",
-      "setpix",
-      "setq",
-      "setq2",
-      "setr",
-      "sets",
-      "setscp",
-      "setse1",
-      "setse2",
-      "setse3",
-      "setse4",
-      "setword",
-      "setxfrq",
-      "seussf",
-      "seussr",
-      "shl",
-      "shr",
-      "signx",
-      "skip",
-      "skipf",
-      "splitb",
-      "splitw",
-      "stalli",
-      "sub",
-      "subr",
-      "subs",
-      "subsx",
-      "subx",
-      "sumc",
-      "sumnc",
-      "sumnz",
-      "sumz",
-      "test",
-      "testb",
-      "testbn",
-      "testn",
-      "testp",
-      "testpn",
-      "tjf",
-      "tjnf",
-      "tjns",
-      "tjnz",
-      "tjs",
-      "tjv",
-      "tjz",
-      "trgint1",
-      "trgint2",
-      "trgint3",
-      "waitatn",
-      "waitct1",
-      "waitct2",
-      "waitct3",
-      "waitfbw",
-      "waitint",
-      "waitpat",
-      "waitse1",
-      "waitse2",
-      "waitse3",
-      "waitse4",
-      "waitx",
-      "waitxfi",
-      "waitxmt",
-      "waitxrl",
-      "waitxro",
-      "wfbyte",
-      "wflong",
-      "wfword",
-      "wmlong",
-      "wrbyte",
-      "wrc",
-      "wrfast",
-      "wrlong",
-      "wrlut",
-      "wrnc",
-      "wrnz",
-      "wrpin",
-      "wrword",
-      "wrz",
-      "wxpin",
-      "wypin",
-      "xcont",
-      "xinit",
-      "xor",
-      "xoro32",
-      "xstop",
-      "xzero",
-      "zerox",
-    ];
-    const instructionStatus: boolean = pasmInstructions.indexOf(name.toLowerCase()) != -1;
-    return instructionStatus;
-  }
-
-  private _isPasmNonArgumentInstruction(name: string): boolean {
-    const pasmNonArgumentInstructions: string[] = [
-      "nop",
-      "resi3",
-      "resi2",
-      "resi1",
-      "resi0",
-      "reti3",
-      "reti2",
-      "reti1",
-      "reti0",
-      "xstop",
-      "allowi",
-      "stalli",
-      "trgint1",
-      "trgint2",
-      "trgint3",
-      "nixint1",
-      "nixint2",
-      "nixint3",
-      "ret",
-      "reta",
-      "retb",
-      "pollint",
-      "pollct1",
-      "pollct2",
-      "pollct3",
-      "pollse1",
-      "pollse2",
-      "pollse3",
-      "pollse4",
-      "pollpat",
-      "pollfbw",
-      "pollxmt",
-      "pollxfi",
-      "pollxro",
-      "pollxrl",
-      "pollatn",
-      "pollqmt",
-      "waitint",
-      "waitct1",
-      "waitct2",
-      "waitct3",
-      "waitse1",
-      "waitse2",
-      "waitse3",
-      "waitse4",
-      "waitpat",
-      "waitfbw",
-      "waitxmt",
-      "waitxfi",
-      "waitxro",
-      "waitxrl",
-      "waitatn",
-    ];
-    const instructionStatus: boolean = pasmNonArgumentInstructions.indexOf(name.toLowerCase()) != -1;
-    return instructionStatus;
-  }
-
-  private _isIllegalInlinePasmDirective(name: string): boolean {
-    const illegalInlinePasmDirective: string[] = ["alignw", "alignl", "file", "orgh"];
-    const illegalStatus: boolean = illegalInlinePasmDirective.indexOf(name.toLowerCase()) != -1;
-    return illegalStatus;
-  }
-
-  private _isPasmConditional(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length >= 2) {
-      const checkType: string = name.toUpperCase();
-      if (
-        checkType == "WC" ||
-        checkType == "WZ" ||
-        checkType == "WCZ" ||
-        checkType == "XORC" ||
-        checkType == "XORZ" ||
-        checkType == "ORC" ||
-        checkType == "ORZ" ||
-        checkType == "ANDC" ||
-        checkType == "ANDZ"
-      ) {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isDatOrPasmLabel(name: string): boolean {
-    let haveLabelStatus: boolean = name.substr(0, 1).match(/[a-zA-Z_\.\:]/) ? true : false;
-    if (haveLabelStatus) {
-      if (this._isDatNFileStorageType(name)) {
-        haveLabelStatus = false;
-      } else if (name.toUpperCase() == "DAT") {
-        haveLabelStatus = false;
-      } else if (this._isReservedPasmSymbols(name)) {
-        haveLabelStatus = false;
-      } else if (name.toUpperCase().startsWith("IF_") || name.toUpperCase() == "_RET_") {
-        haveLabelStatus = false;
-      } else if (this._isPasmConditional(name)) {
-        haveLabelStatus = false;
-      } else if (this._isIllegalInlinePasmDirective(name)) {
-        haveLabelStatus = false;
-      } else if (this._isPasmInstruction(name)) {
-        haveLabelStatus = false;
-      } else if (this._isPasmNonArgumentInstruction(name)) {
-        haveLabelStatus = false;
-      }
-    }
-    return haveLabelStatus;
-  }
-
-  private _isDatNFileStorageType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 2) {
-      const checkType: string = name.toUpperCase();
-      // yeah, FILE too!  (oddly enough)
-      if (checkType == "FILE") {
-        returnStatus = true;
-      } else {
-        returnStatus = this._isDatStorageType(name);
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isDatStorageType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 2) {
-      const checkType: string = name.toUpperCase();
-      if (checkType == "RES") {
-        returnStatus = true;
-      } else {
-        returnStatus = this._isStorageType(name);
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isStorageType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 3) {
-      const checkType: string = name.toUpperCase();
-      if (checkType == "BYTEFIT" || checkType == "WORDFIT" || checkType == "BYTE" || checkType == "WORD" || checkType == "LONG") {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  private _isAlignType(name: string): boolean {
-    let returnStatus: boolean = false;
-    if (name.length > 5) {
-      const checkType: string = name.toUpperCase();
-      if (checkType == "ALIGNL" || checkType == "ALIGNW") {
-        returnStatus = true;
-      }
-    }
-    return returnStatus;
-  }
-
-  // ---------------- new deug() support ----------------
-  //  updated: 26 Mar 2022  - (as of Spin2 v35s)
-  //  debug() statements for special displays support the following
-  //    plot      - General-purpose plotter with cartesian and polar modes
-  //    term      - Text terminal with up to 300 x 200 characters, 6..200 point font size, 4 simultaneous color schemes
-  //    midi      - Piano keyboard with 1..128 keys, velocity depiction, variable screen scale
-  //    logic     - PDM, Logic analyzer with single and multi-bit labels, 1..32 channels, can trigger on pattern
-  //    scope     - PDM, Oscilloscope with 1..8 channels, can trigger on level with hysteresis
-  //    scope_xy  - PDM, XY oscilloscope with 1..8 channels, persistence of 0..512 samples, polar mode, log scale mode
-  //    fft       - PDM, Fast Fourier Transform with 1..8 channels, 4..2048 points, windowed results, log scale mode
-  //    spectro   - PDM, Spectrograph with 4..2048-point FFT, windowed results, phase-coloring, and log scale mode
-  //    bitmap    - PDM, Bitmap, 1..2048 x 1..2048 pixels, 1/2/4/8/16/32-bit pixels with 19 color systems, 15 direction/autoscroll modes, independent X and Y pixel size of 1..256
-  // ----------------------------------------------------
-  private _isDebugDisplayType(name: string): boolean {
-    const debugDisplayTypes: string[] = ["logic", "scope", "scope_xy", "fft", "spectro", "plot", "term", "bitmap", "midi"];
-    //const bDisplayTypeStatus: boolean = (debugDisplayTypes.indexOf(name.toLowerCase()) != -1);
-    const bDisplayTypeStatus: boolean = debugDisplayTypes.includes(name.toLowerCase());
-    return bDisplayTypeStatus;
-  }
-
-  private _isNameWithTypeInstantiation(newParameter: string, displayType: eDebugDisplayType): boolean {
-    var nameStatus: boolean = false;
-    const bHasPackedData: boolean = this._typeHasPackedData(displayType);
-    const bHasColorMode: boolean = this._typeHasColorMode(displayType);
-    switch (displayType) {
-      case eDebugDisplayType.ddtTerm:
-        nameStatus = this._isDebugTermDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtScope:
-        nameStatus = this._isDebugScopeDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtScopeXY:
-        nameStatus = this._isDebugScopeXYDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtLogic:
-        nameStatus = this._isDebugLogicDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtFFT:
-        nameStatus = this._isDebugFFTDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtSpectro:
-        nameStatus = this._isDebugSpectroDeclarationParam(newParameter);
-        // SPECTRO-Instantiation supports a special color mode, check it too
-        if (nameStatus == false) {
-          nameStatus = this._isDebugSpectroColorMode(newParameter);
-        }
-        break;
-      case eDebugDisplayType.ddtPlot:
-        nameStatus = this._isDebugPlotDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtBitmap:
-        nameStatus = this._isDebugBitmapDeclarationParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtMidi:
-        nameStatus = this._isDebugMidiDeclarationParam(newParameter);
-        break;
-      default:
-        break;
-    }
-    // if we don't have a match yet then check packed data
-    if (nameStatus == false && bHasPackedData) {
-      nameStatus = this._isDebugPackedDataType(newParameter);
-    }
-    if (nameStatus == false && bHasColorMode) {
-      nameStatus = this._isDebugBitmapColorMode(newParameter);
-    }
-    return nameStatus;
-  }
-
-  private _isNameWithTypeFeed(newParameter: string, displayType: eDebugDisplayType): boolean {
-    var nameStatus: boolean = false;
-    const bHasColorMode: boolean = this._typeHasColorMode(displayType);
-    switch (displayType) {
-      case eDebugDisplayType.ddtTerm:
-        nameStatus = this._isDebugTermFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtScope:
-        nameStatus = this._isDebugScopeFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtScopeXY:
-        nameStatus = this._isDebugScopeXYFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtLogic:
-        nameStatus = this._isDebugLogicFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtFFT:
-        nameStatus = this._isDebugFFTFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtSpectro:
-        nameStatus = this._isDebugSpectroFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtPlot:
-        nameStatus = this._isDebugPlotFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtBitmap:
-        nameStatus = this._isDebugBitmapFeedParam(newParameter);
-        break;
-      case eDebugDisplayType.ddtMidi:
-        nameStatus = this._isDebugMidiFeedParam(newParameter);
-        break;
-      default:
-        break;
-    }
-    // if we don't have a match yet then check color mode
-    if (nameStatus == false && bHasColorMode) {
-      nameStatus = this._isDebugBitmapColorMode(newParameter);
-    }
-    this._logDEBUG("  -- _isNameWithTypeFeed(" + newParameter + ", " + displayType + ") = " + nameStatus);
-    return nameStatus;
-  }
-
-  // each type has decl and feed parameter-name check methods
-  // Debug Display: TERM declaration
-  private _isDebugTermDeclarationParam(name: string): boolean {
-    const debugTermDeclTypes: string[] = ["title", "pos", "size", "textsize", "color", "backcolor", "update", "hidexy"];
-    const bTermDeclParamStatus: boolean = debugTermDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bTermDeclParamStatus;
-  }
-
-  // Debug Display: TERM feed
-  private _isDebugTermFeedParam(name: string): boolean {
-    const debugTermFeedTypes: string[] = ["clear", "update", "save", "close"];
-    const bTermFeedParamStatus: boolean = debugTermFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bTermFeedParamStatus;
-  }
-
-  // Debug Display: SCOPE declaration
-  private _isDebugScopeDeclarationParam(name: string): boolean {
-    const debugScopeDeclTypes: string[] = ["title", "pos", "size", "samples", "rate", "dotsize", "linesize", "textsize", "color", "hidexy"];
-    const bScopeDeclParamStatus: boolean = debugScopeDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bScopeDeclParamStatus;
-  }
-
-  // Debug Display: SCOPE feed
-  private _isDebugScopeFeedParam(name: string): boolean {
-    const debugScopeFeedTypes: string[] = ["trigger", "holdoff", "samples", "clear", "save", "window", "close"];
-    const bScopeFeedParamStatus: boolean = debugScopeFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bScopeFeedParamStatus;
-  }
-
-  // Debug Display: SCOPE_XY declaration
-  private _isDebugScopeXYDeclarationParam(name: string): boolean {
-    const debugScopeXYDeclTypes: string[] = ["title", "pos", "size", "range", "samples", "rate", "dotsize", "textsize", "color", "polar", "logscale", "hidexy"];
-    const bScopeXYDeclParamStatus: boolean = debugScopeXYDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bScopeXYDeclParamStatus;
-  }
-
-  // Debug Display: SCOPE_XY feed
-  private _isDebugScopeXYFeedParam(name: string): boolean {
-    const debugScopeXYFeedTypes: string[] = ["clear", "save", "window", "close"];
-    const bScopeXYFeedParamStatus: boolean = debugScopeXYFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bScopeXYFeedParamStatus;
-  }
-
-  // Debug Display: LOGIC declaration
-  private _isDebugLogicDeclarationParam(name: string): boolean {
-    const debugLogicDeclTypes: string[] = ["title", "pos", "samples", "spacing", "rate", "linesize", "textsize", "color", "hidexy"];
-    const bLogicDeclParamStatus: boolean = debugLogicDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bLogicDeclParamStatus;
-  }
-
-  // Debug Display: LOGIC feed
-  private _isDebugLogicFeedParam(name: string): boolean {
-    const debugLogicFeedTypes: string[] = ["trigger", "holdoff", "clear", "save", "window", "close"];
-    const bLogicFeedParamStatus: boolean = debugLogicFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bLogicFeedParamStatus;
-  }
-
-  // Debug Display: FFT declaration
-  private _isDebugFFTDeclarationParam(name: string): boolean {
-    const debugFFTDeclTypes: string[] = ["title", "pos", "size", "samples", "rate", "dotsize", "linesize", "textsize", "color", "logscale", "hidexy"];
-    const bFFTDeclParamStatus: boolean = debugFFTDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bFFTDeclParamStatus;
-  }
-
-  // Debug Display: FFT feed
-  private _isDebugFFTFeedParam(name: string): boolean {
-    const debugFFTFeedTypes: string[] = ["clear", "save", "window", "close"];
-    const bFFTFeedParamStatus: boolean = debugFFTFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bFFTFeedParamStatus;
-  }
-
-  // Debug Display: SPECTRO declaration
-  private _isDebugSpectroDeclarationParam(name: string): boolean {
-    const debugSpectroDeclTypes: string[] = ["title", "pos", "samples", "depth", "mag", "range", "rate", "trace", "dotsize", "logscale", "hidexy"];
-    const bSpectroDeclParamStatus: boolean = debugSpectroDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bSpectroDeclParamStatus;
-  }
-
-  // Debug Display: SPECTRO feed
-  private _isDebugSpectroFeedParam(name: string): boolean {
-    const debugSpectroFeedTypes: string[] = ["clear", "save", "window", "close"];
-    const bSpectroFeedParamStatus: boolean = debugSpectroFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bSpectroFeedParamStatus;
-  }
-
-  // Debug Display: PLOT declaration
-  private _isDebugPlotDeclarationParam(name: string): boolean {
-    const debugPlotDeclTypes: string[] = ["title", "pos", "size", "dotsize", "lutcolors", "backcolor", "update", "hidexy"];
-    const bPlotDeclParamStatus: boolean = debugPlotDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bPlotDeclParamStatus;
-  }
-
-  // Debug Display: PLOT feed
-  private _isDebugPlotFeedParam(name: string): boolean {
-    const debugPlotFeedTypes: string[] = [
-      "lutcolors",
-      "backcolor",
-      "color",
-      "opacity",
-      "precise",
-      "linesize",
-      "origin",
-      "set",
-      "dot",
-      "line",
-      "circle",
-      "oval",
-      "box",
-      "obox",
-      "textsize",
-      "textstyle",
-      "textangle",
-      "text",
-      "spritedef",
-      "sprite",
-      "polar",
-      "cartesian",
-      "update",
-      "clear",
-      "save",
-      "window",
-      "close",
-    ];
-    const bPlotFeedParamStatus: boolean = debugPlotFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bPlotFeedParamStatus;
-  }
-
-  // Debug Display: BITMAP declaration
-  private _isDebugBitmapDeclarationParam(name: string): boolean {
-    const debugBitmapDeclTypes: string[] = ["title", "pos", "size", "dotsize", "lutcolors", "trace", "rate", "update", "hidexy", "sparse"];
-    const bBitmapDeclParamStatus: boolean = debugBitmapDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bBitmapDeclParamStatus;
-  }
-
-  // Debug Display: BITMAP feed
-  private _isDebugBitmapFeedParam(name: string): boolean {
-    const debugBitmapFeedTypes: string[] = ["lutcolors", "trace", "rate", "set", "scroll", "clear", "update", "save", "window", "close"];
-    const bBitmapFeedParamStatus: boolean = debugBitmapFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bBitmapFeedParamStatus;
-  }
-
-  // Debug Display: MIDI declaration
-  private _isDebugMidiDeclarationParam(name: string): boolean {
-    const debugMidiDeclTypes: string[] = ["title", "pos", "size", "range", "channel", "color"];
-    const bMidiDeclParamStatus: boolean = debugMidiDeclTypes.indexOf(name.toLowerCase()) != -1;
-    return bMidiDeclParamStatus;
-  }
-
-  // Debug Display: MIDI feed
-  private _isDebugMidiFeedParam(name: string): boolean {
-    const debugMidiFeedTypes: string[] = ["clear", "save", "window", "close"];
-    const bMidiFeedParamStatus: boolean = debugMidiFeedTypes.indexOf(name.toLowerCase()) != -1;
-    return bMidiFeedParamStatus;
-  }
-
-  private _typeHasPackedData(displayType: eDebugDisplayType): boolean {
-    // return indication if displayType has Packed Data Mode
-    let bHasPackedData: boolean = true;
-    switch (displayType) {
-      case eDebugDisplayType.ddtTerm:
-        bHasPackedData = false;
-        break;
-      case eDebugDisplayType.ddtPlot:
-        bHasPackedData = false;
-        break;
-      case eDebugDisplayType.ddtMidi:
-        bHasPackedData = false;
-        break;
-      default:
-        break;
-    }
-    return bHasPackedData;
-  }
-
-  private _typeHasColorMode(displayType: eDebugDisplayType): boolean {
-    // return indication if displayType has lut1_to_rgb24 Color Mode
-    let bHasColorMode: boolean = false;
-    switch (displayType) {
-      case eDebugDisplayType.ddtBitmap:
-        bHasColorMode = true;
-        break;
-      case eDebugDisplayType.ddtPlot:
-        bHasColorMode = true;
-        break;
-      default:
-        break;
-    }
-    return bHasColorMode;
-  }
-
-  // color names for use in debug()
-  //   BLACK / WHITE or ORANGE / BLUE / GREEN / CYAN / RED / MAGENTA / YELLOW / GREY|GRAY
-  private _isDebugColorName(name: string): boolean {
-    const debugColorNames: string[] = ["black", "white", "orange", "blue", "green", "cyan", "red", "magenta", "yellow", "grey", "gray"];
-    const bColorNameStatus: boolean = debugColorNames.indexOf(name.toLowerCase()) != -1;
-    return bColorNameStatus;
-  }
-
-  // packed data forms for use in debug()
-  private _isDebugPackedDataType(name: string): boolean {
-    const debugPackedDataTypes: string[] = [
-      "longs_1bit",
-      "longs_2bit",
-      "longs_4bit",
-      "longs_8bit",
-      "longs_16bit",
-      "words_1bit",
-      "words_2bit",
-      "words_4bit",
-      "words_8bit",
-      "bytes_1bit",
-      "bytes_2bit",
-      "bytes_4bit",
-      // optional operators
-      "alt",
-      "signed",
-    ];
-    const bPackedDataTypeStatus: boolean = debugPackedDataTypes.indexOf(name.toLowerCase()) != -1;
-    return bPackedDataTypeStatus;
-  }
-
-  //  Bitmap Color Modes
-  private _isDebugBitmapColorMode(name: string): boolean {
-    const debugBitmapColorModes: string[] = [
-      "lut1",
-      "lut2",
-      "lut4",
-      "lut8",
-      "luma8",
-      "luma8w",
-      "luma8x",
-      "hsv8",
-      "hsv8w",
-      "hsv8x",
-      "rgbi8",
-      "rgbi8w",
-      "rgbi8x",
-      "rgb8",
-      "rgb16",
-      "rgb24",
-      "hsv16",
-      "hsv16w",
-      "hsv16x",
-    ];
-    const bBitmapColorModeStatus: boolean = debugBitmapColorModes.indexOf(name.toLowerCase()) != -1;
-    return bBitmapColorModeStatus;
-  }
-
-  //  Spectro reduced-set Color Modes
-  private _isDebugSpectroColorMode(name: string): boolean {
-    const debugSpectropColorModes: string[] = ["luma8", "luma8w", "luma8x", "hsv16", "hsv16w", "hsv16x"];
-    const bSpectroColorModeStatus: boolean = debugSpectropColorModes.indexOf(name.toLowerCase()) != -1;
-    return bSpectroColorModeStatus;
-  }
-
   //
   // ----------------------------------------------------
-
-  private _skipWhite(line: string, currentOffset: number): number {
-    let firstNonWhiteIndex: number = currentOffset;
-    for (let index = currentOffset; index < line.length; index++) {
-      if (line.substr(index, 1) != " " && line.substr(index, 1) != "\t") {
-        firstNonWhiteIndex = index;
-        break;
-      }
-    }
-    return firstNonWhiteIndex;
-  }
 
   private _tokenString(aToken: IParsedToken, line: string): string {
     let varName: string = line.substr(aToken.startCharacter, aToken.length);
